% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCapacity.R
\name{calcCapacity}
\alias{calcCapacity}
\title{calcCapacity}
\usage{
calcCapacity()
}
\value{
The ENERDATA data filtered by Capacity.
}
\description{
Derive data on capacity from ENERDATA.
}
\examples{
\dontrun{
a <- calcOutput(type = "Capacity", file = "iCapacity.csvr")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
