% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCO2_emissions.R
\name{calcCO2_emissions}
\alias{calcCO2_emissions}
\title{calcCO2_emissions}
\usage{
calcCO2_emissions()
}
\value{
The read-in data into a magpie object.
}
\description{
Derive CO2_emissions data based on data sources:
IEA_CO2 for period 2021
EDGAR for period 1970:2022
The final CO2 emissions is calculated by finding the share between
energy combustion and industrial process emissions from IEA_CO2 and multiply
the share with EDGAR data.
}
\examples{
\dontrun{
a <- calcOutput(type = "CO2_emissions", aggregate = TRUE)
}
}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
