% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConstantPrice.R
\name{calcConstantPrice}
\alias{calcConstantPrice}
\title{calcConstantPrice}
\usage{
calcConstantPrice()
}
\value{
The "ENERDATA" data filtered by constant price.
}
\description{
Read in several excel files with data from ENERDATA.
}
\details{
The two available ENERDATA datasets are combined into one,
NA values are replaced with zeros and historical data are corrected.

Finally, the combined ENERDATA datasets are filtered by constant price.
}
\examples{
\dontrun{
a <- calcOutput(type = "ConstantPrice", file = "ConstantPrice.csv" , aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
