% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEMBERCapacity.R
\name{calcEMBERCapacity}
\alias{calcEMBERCapacity}
\title{calcEMBERCapacity}
\usage{
calcEMBERCapacity()
}
\value{
magpie object
}
\description{
Use data from EMBER Capacity, in GW.
}
\examples{
\dontrun{
a <- calcOutput(type = "EMBERCapacity", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
