% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEMBERElecProd.R
\name{calcEMBERElecProd}
\alias{calcEMBERElecProd}
\title{calcEMBERElecProd}
\usage{
calcEMBERElecProd()
}
\value{
magpie object
}
\description{
Use data from EMBER Electricity generation, in TWh.
}
\examples{
\dontrun{
a <- calcOutput(type = "EMBERElecProd", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
