% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnergyCosts.R
\name{calcEnergyCosts}
\alias{calcEnergyCosts}
\title{calcEnergyCosts}
\usage{
calcEnergyCosts(subtype = "PowerAndHeat")
}
\arguments{
\item{subtype}{Type of data that should be read. The type is referring to the
excel sheet, from the excel file "REF2020_Technology Assumptions_Energy.xlsx"
Available types are:
\itemize{
\item `PowerAndHeat`:
\item `DomesticEnergy`:
\item `IndustryEnergy`:
\item `new_fuels_energy`:
\item `renovation_costs`:
}}
}
\value{
Read EnergyCosts data and convert it to a csv file
}
\description{
Read EnergyCosts data from the "EU Reference Scenario":
https://energy.ec.europa.eu/data-and-analysis/energy-modelling/eu-reference-scenario-2020_en
and convert it to a csv file
}
\examples{
\dontrun{
a <- calcOutput(type = "EnergyCosts", file = "PowerAndHeat.csv", aggregate = FALSE)
}
}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
