% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEurostat2.R
\name{calcEurostat2}
\alias{calcEurostat2}
\title{calcEurostat2}
\usage{
calcEurostat2(subtype = "SIOT", subset = "all", select = "one")
}
\arguments{
\item{subtype}{Type of data.
Available types are:
\itemize{
\item `SIOT`:Symmetric input-output table at basic prices (product by product)
\item `SUPL`:Supply table at basic prices incl. transformation into purchasers' prices
\item `USEp`:Use table at purchasers' prices
\item `TRM`:Table of trade and transport margins
\item `TAX`:Table of taxes less subsidies on product
\item `BTRADE`:International trade in services
\item `POP_NAT`:Population and employment
\item `EMP_SEC`:National accounts aggregates by industry
\item `DEMO`:Population on 1 January by age and sex
\item `LFS_POP`:Population by sex, age, citizenship and labour status
\item `LFS_EMP`:Employment by sex, age, occupation and economic activity
\item `LFS_UNEMP`:Previous occupations of the unemployed, by sex
\item `NFTR`:Non-financial transactions - annual data
\item `FTR`:Financial transactions - annual data
\item `ENV`:Air emissions accounts by NACE Rev. 2 activity
\item `COICOP`:Final consumption expenditure of households by consumption purpose
\item `ALL`:All the available subtypes
}}

\item{subset}{Read part of the data.}

\item{select}{Select one or all the available subtypes.
\itemize{
\item `one`:One subtype
\item `all`:All the available subtypes
}}
}
\value{
Eurostat data into a gdx file
}
\description{
calcEurostat2
}
\examples{
\dontrun{
calcOutput("Eurostat2")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
