% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFE.R
\name{calcFE}
\alias{calcFE}
\title{calcFE}
\usage{
calcFE(source = "IEA", scenario_proj = "SSP2")
}
\arguments{
\item{source}{"IEA"}

\item{scenario_proj}{"SSP2" by default unless overwritten}
}
\value{
Final Energy data from IEA data as MAgPIE object aggregated to country level
}
\description{
Calculates FE historical from IEA energy balances
}
\examples{
\dontrun{
a <- calcOutput("FE")
}

}
\author{
Lavinia Baumstark, Aman Malik
}
