% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcfullACTV.R
\name{calcfullACTV}
\alias{calcfullACTV}
\title{calcfullACTV}
\usage{
calcfullACTV()
}
\value{
The read-in data into a magpie object.
}
\description{
Derive economic activity data for OPENPROM sectors based on two data sources:
transport, traffic, air transport passengers per country and per year (IRF)
and Production Level and Unit Cost (GEME3).
}
\examples{
\dontrun{
a <- calcOutput(type = "fullACTV", aggregate = TRUE)
}
}
\author{
Fotis Sioutas
}
