% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHydrogenDemand.R
\name{calcHydrogenDemand}
\alias{calcHydrogenDemand}
\title{calcHydrogenDemand}
\usage{
calcHydrogenDemand()
}
\value{
HydrogenDemand fuel consumption
}
\description{
Use HydrogenDemand data for consumption
}
\examples{
\dontrun{
a <- calcOutput(type = "HydrogenDemand", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
