% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIACTV_OP_GEM.R
\name{calcIACTV_OP_GEM}
\alias{calcIACTV_OP_GEM}
\title{calcIACTV_OP_GEM}
\usage{
calcIACTV_OP_GEM()
}
\value{
The read-in data into a magpie object.
}
\description{
calcIACTV_OP_GEM
}
\examples{
\dontrun{
a <- calcOutput(type = "IACTV_OP_GEM", aggregate = TRUE)
}
}
\author{
Fotis Sioutas
}
