% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIChpPowGen.R
\name{calcIChpPowGen}
\alias{calcIChpPowGen}
\title{calcIChpPowGen}
\usage{
calcIChpPowGen()
}
\value{
magpie object with OPENPROM input data iChpPowGen
}
\description{
Combined Heat and Power (CHP) and District Heating Power Generation
(DHP) techno-economic parameters for electricity production are derived
using data from the EU Reference Scenario and technology cost assumptions
from TechCosts2024(EU Reference Scenario). The dataset includes investment costs (IC),
fixed costs (FC), variable operation and maintenance costs (VOM), plant
lifetime (LFT), thermal efficiency (effThrm), and technology-specific
availability factors (AVAIL) for CHP electricity generation technologies.
Technology cost values are converted to OPEN-PROM monetary and energy
units where necessary.
Technologies are mapped from PRIMES and TechCosts classifications to OPEN-PROM CHP
technology categories and technology-specific availability assumptions
are hardcoded.
Missing years are interpolated over the full model horizon.
}
\examples{
\dontrun{
a <- calcOutput(type = "IChpPowGen", aggregate = FALSE)
}

}
\author{
Alexandros, Michael Madianos
}
