% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataAggregateService.R
\name{calcIDataAggregateService}
\alias{calcIDataAggregateService}
\title{calcIDataAggregateService}
\usage{
calcIDataAggregateService()
}
\value{
Magpie object with the FAOProductionCrops
}
\description{
calcIDataAggregateService
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataAggregateService", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
