% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataAgricultureEff.R
\name{calcIDataAgricultureEff}
\alias{calcIDataAgricultureEff}
\title{calcIDataAgricultureEff}
\usage{
calcIDataAgricultureEff()
}
\value{
Magpie object with the FAOProductionCrops
}
\description{
calcIDataAgricultureEff
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataAgricultureEff", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
