% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataAgriculturePrice.R
\name{calcIDataAgriculturePrice}
\alias{calcIDataAgriculturePrice}
\title{calcIDataAgriculturePrice}
\usage{
calcIDataAgriculturePrice()
}
\value{
Magpie object with Agriculture Price
}
\description{
calcIDataAgriculturePrice
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataAgriculturePrice", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
