% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataAgricultureService.R
\name{calcIDataAgricultureService}
\alias{calcIDataAgricultureService}
\title{calcIDataAgricultureService}
\usage{
calcIDataAgricultureService()
}
\value{
Magpie object with the FAOProductionCrops
}
\description{
calcIDataAgricultureService
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataAgricultureService", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
