% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataAgricultureTFC.R
\name{calcIDataAgricultureTFC}
\alias{calcIDataAgricultureTFC}
\title{calcIDataAgricultureTFC}
\usage{
calcIDataAgricultureTFC()
}
\value{
OPENPROM historical shares of agriculture.
}
\description{
Use IFuelCons and data from AGENRES to dissagregate the consumption
in various agriculture modes.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataAgricultureTFC", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
