% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataCalibUsefulEnergy.R
\name{calcIDataCalibUsefulEnergy}
\alias{calcIDataCalibUsefulEnergy}
\title{calcIDataCalibUsefulEnergy}
\usage{
calcIDataCalibUsefulEnergy()
}
\value{
magpie object with OPENPROM input data iCalibUsefulEnergy
}
\description{
Use uncalibrated data to derive default values for iCalibUsefulEnergy
It includes calibration parameters for useful energy in the industry module.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataCalibUsefulEnergy", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
