% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataCaloriesIntake.R
\name{calcIDataCaloriesIntake}
\alias{calcIDataCaloriesIntake}
\title{calcIDataCaloriesIntake}
\usage{
calcIDataCaloriesIntake()
}
\value{
Magpie object with the Food supply (kcal/capita/day)
}
\description{
calcIDataCaloriesIntake
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataCaloriesIntake", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
