% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataConsEneBranch.R
\name{calcIDataConsEneBranch}
\alias{calcIDataConsEneBranch}
\title{calcIDataConsEneBranch}
\usage{
calcIDataConsEneBranch()
}
\value{
OPENPROM input data iDataConsEneBranch
}
\description{
Use data to derive OPENPROM input parameter iDataConsEneBranch
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataConsEneBranch", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
