% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataDistrLosses.R
\name{calcIDataDistrLosses}
\alias{calcIDataDistrLosses}
\title{calcIDataDistrLosses}
\usage{
calcIDataDistrLosses()
}
\value{
OPENPROM input data iDataDistrLosses
}
\description{
This function processes IEA distribution loss data, converts units to Mtoe,
maps fuels to OPENPROM sectors, aggregates across regions and periods,
and fills missing country values with zeros.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataDistrLosses", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
