% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataElecAndSteamGen.R
\name{calcIDataElecAndSteamGen}
\alias{calcIDataElecAndSteamGen}
\title{calcIDataElecAndSteamGen}
\usage{
calcIDataElecAndSteamGen()
}
\value{
OPENPROM input data iDataElecAndSteamGen
}
\description{
Use data to derive OPENPROM input parameter iDataElecAndSteamGen
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataElecAndSteamGen", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
