% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataElecSteamGen.R
\name{calcIDataElecSteamGen}
\alias{calcIDataElecSteamGen}
\title{calcIDataElecSteamGen}
\usage{
calcIDataElecSteamGen()
}
\value{
OPENPROM input data iDataElecSteamGen
}
\description{
Use data to derive OPENPROM input parameter iDataElecSteamGen
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataElecSteamGen", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
