% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataHeatProd.R
\name{calcIDataHeatProd}
\alias{calcIDataHeatProd}
\title{calcIDataHeatProd}
\usage{
calcIDataHeatProd()
}
\value{
OPENPROM input data iDataHeatProd
}
\description{
Derives the OPENPROM input parameter {IDataHeatProd} from
IEA heat production statistics.
The function processes IEA heat production data for combined heat and power
(CHP) and steam/heat production activities using the following IEA flows:
{HEMAINC} and {HEAUTOC} for CHP heat production
{HEMAINH} and {HEAUTOH} for steam and district heat production
IEA fuel products are mapped to OPENPROM energy forms (EFs), which are
subsequently linked to OPENPROM CHP and district heating technologies using
predefined mapping tables.
Heat production values are converted from KTOE to Mtoe and aggregated by
region, period, and technology.
The resulting dataset provides heat production data in Mtoe for use as
OPENPROM input data.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataHeatProd", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
