% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataImports.R
\name{calcIDataImports}
\alias{calcIDataImports}
\title{calcIDataImports}
\usage{
calcIDataImports()
}
\value{
magpie object with OPENPROM input data iDataImports.
}
\description{
Use data from ENERDATA to derive OPENPROM input parameter iDataImports.
This dataset includes imports values for each region and fuel type in Mtoe.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataImports", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
