% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataPassCars.R
\name{calcIDataPassCars}
\alias{calcIDataPassCars}
\title{calcIDataPassCars}
\usage{
calcIDataPassCars()
}
\value{
OPENPROM input data iDataPassCars
}
\description{
Use data to derive OPENPROM input parameter iDataPassCars
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataPassCars", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
