% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataPGScaleEndogScrap.R
\name{calcIDataPGScaleEndogScrap}
\alias{calcIDataPGScaleEndogScrap}
\title{calcIDataPGScaleEndogScrap}
\usage{
calcIDataPGScaleEndogScrap()
}
\value{
magpie object with OPENPROM input data iScaleEndogScrap
}
\description{
Use uncalibrated data to derive default values for iScaleEndogScrap
It includes parameters for premature scrapping in the PG module.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataPGScaleEndogScrap", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
