% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataPremScrpFac.R
\name{calcIDataPremScrpFac}
\alias{calcIDataPremScrpFac}
\title{calcIDataPremScrpFac}
\usage{
calcIDataPremScrpFac()
}
\value{
magpie object with OPENPROM input data iPremScrpFac
}
\description{
Use uncalibrated data to derive default values for iPremScrpFac
This dataset includes Maturity factors for premature scrapping .
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataPremScrpFac", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
