% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataScaleEndogScrap.R,
%   R/calcIDataScaleEndogScrapPG.R
\name{calcIDataScaleEndogScrap}
\alias{calcIDataScaleEndogScrap}
\title{calcIDataScaleEndogScrap}
\usage{
calcIDataScaleEndogScrap()

calcIDataScaleEndogScrap()
}
\value{
magpie object with OPENPROM input data iScaleEndogScrap

magpie object with OPENPROM input data iScaleEndogScrap
}
\description{
Use uncalibrated data to derive default values for iScaleEndogScrap
It includes parameters for premature scrapping in the industry module.

Use uncalibrated data to derive default values for iScaleEndogScrap
It includes parameters for premature scrapping in the PG module.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataScaleEndogScrap", aggregate = FALSE)
}

\dontrun{
a <- calcOutput(type = "IDataScaleEndogScrapPG", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
