% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataSharesAgriculture.R
\name{calcIDataSharesAgriculture}
\alias{calcIDataSharesAgriculture}
\title{calcIDataSharesAgriculture}
\usage{
calcIDataSharesAgriculture()
}
\value{
OPENPROM historical shares of agriculture.
}
\description{
Use IFuelCons and data from AGENRES to dissagregate the consumption
in various agriculture modes.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataSharesAgriculture", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
