% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataTotTransfInputRef.R
\name{calcIDataTotTransfInputRef}
\alias{calcIDataTotTransfInputRef}
\title{calcIDataTotTransfInputRef}
\usage{
calcIDataTotTransfInputRef()
}
\value{
OPENPROM input data iDataTotTransfInputRef.
}
\description{
Use data from ENERDATA to derive OPENPROM input parameter iDataTotTransfInputRef.
This dataset has the refineries input in Mtoe.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataTotTransfInputRef", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
