% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataTransfOutputRef.R
\name{calcIDataTransfOutputRef}
\alias{calcIDataTransfOutputRef}
\title{calcIDataTransfOutputRef}
\usage{
calcIDataTransfOutputRef()
}
\value{
magpie object with OPENPROM input data iDataImports.
}
\description{
Use data from ENERDATA to derive OPENPROM input parameter iDataTransfOutputRef.
This dataset includes the transformation Output from Refineries.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataTransfOutputRef", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
