% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDomFuelConsPerSub.R
\name{calcIDomFuelConsPerSub}
\alias{calcIDomFuelConsPerSub}
\title{calcIDomFuelConsPerSub}
\usage{
calcIDomFuelConsPerSub()
}
\value{
OPENPROM input data iDomFuelConsPerSub
}
\description{
Use ENERDATA fuel consumption data to derive OPENPROM input parameter iDomFuelConsPerSub
(fuel consumption in DOMSE sectors, i.e. SE,AG,HOU)
}
\examples{
\dontrun{
a <- calcOutput(type = "IDomFuelConsPerSub", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
