% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIECEMF.R
\name{calcIECEMF}
\alias{calcIECEMF}
\title{calcIECEMF}
\usage{
calcIECEMF()
}
\value{
The read-in data into a magpie object.
}
\description{
calcIECEMF
}
\examples{
\dontrun{
a <- calcOutput(type = "IECEMF", aggregate = TRUE)
}
}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
