% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIElcNetImpShare.R
\name{calcIElcNetImpShare}
\alias{calcIElcNetImpShare}
\title{calcIElcNetImpShare}
\usage{
calcIElcNetImpShare()
}
\value{
magpie object with OPENPROM input data iSuppRefCapacity.
}
\description{
Use data from ENERDATA to derive OPENPROM input parameter iElcNetImpShare.
This dataset includes the ratio of electricity imports in total final demand.
}
\examples{
\dontrun{
a <- calcOutput(type = "IElcNetImpShare", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
