% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIFinConsSubFuel.R
\name{calcIFinConsSubFuel}
\alias{calcIFinConsSubFuel}
\title{calcIFinConsSubFuel}
\usage{
calcIFinConsSubFuel()
}
\value{
OPENPROM input data iFinConsSubFuel
}
\description{
Use calcFuelConsumption to derive OPENPROM input data iFinConsSubFuel
(fuel consumption in NENSE sectors, i.e. PCH,NEN,BU - non energy and bunkers)
}
\examples{
\dontrun{
a <- calcOutput(type = "IFinConsSubFuel", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
