% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIFuelConsICT.R
\name{calcIFuelConsICT}
\alias{calcIFuelConsICT}
\title{calcIFuelConsICT}
\usage{
calcIFuelConsICT()
}
\value{
Consumption of data centers by scenario and variable.
}
\description{
calcIFuelConsICT
}
\examples{
\dontrun{
a <- calcOutput("IFuelConsICT", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
