% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIH2InfrCapCosts.R
\name{calcIH2IH2InfrCapCosts}
\alias{calcIH2IH2InfrCapCosts}
\title{calcIH2InfrCapCosts}
\usage{
calcIH2IH2InfrCapCosts()
}
\value{
OPENPROM input data H2production.
The output data for Investment cost(IC), Fixed Costs (FC), (EFF),
Variable Costs (VC), (AVAIL), (SELF) and (H2KMTOE) is from excel Common_DATA
from MENA_EDS model.
}
\description{
Use technology cost data from excel Common_DATA from MENA_EDS model
to derive OPENPROM input parameter IH2InfrCapCosts.
}
\examples{
\dontrun{
a <- calcOutput(type = "IH2InfrCapCosts", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
