% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcH2InfrastructureCapCosts.R
\name{calcIH2InfrastructureCapCosts}
\alias{calcIH2InfrastructureCapCosts}
\title{calcIH2InfrastructureCapCosts}
\usage{
calcIH2InfrastructureCapCosts()
}
\value{
OPENPROM input data H2production.
The output data for Investment cost(IC), Fixed Costs (FC), (EFF),
Variable Costs (VC), (AVAIL), (SELF) and (H2KMTOE) is from excel Common_DATA
from MENA_EDS model.
}
\description{
Use technology cost data from excel Common_DATA from MENA_EDS model
to derive OPENPROM input parameter IH2InfrastructureCapCosts.
}
\examples{
\dontrun{
a <- calcOutput(type = "IH2InfrastructureCapCosts", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
