% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIHeatpumpMix.R
\name{calcIHeatpumpMix}
\alias{calcIHeatpumpMix}
\title{calcIHeatpumpMix}
\usage{
calcIHeatpumpMix()
}
\value{
The read-in data into a magpie object.
}
\description{
calcIHeatpumpMix
}
\examples{
\dontrun{
a <- calcOutput(type = "IHeatpumpMix", aggregate = TRUE)
}
}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
