% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIInitSpecFuelCons.R
\name{calcIInitSpecFuelCons}
\alias{calcIInitSpecFuelCons}
\title{calcIInitSpecFuelCons}
\usage{
calcIInitSpecFuelCons()
}
\value{
OPENPROM input data iInitSpecFuelCons.
}
\description{
Use ENERDATA and IRF data to derive OPENPROM input parameter 
iInitSpecFuelCons.
}
\examples{
\dontrun{
a <- calcOutput(type = "IInitSpecFuelCons", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
