% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIInpTransfTherm.R
\name{calcIInpTransfTherm}
\alias{calcIInpTransfTherm}
\title{calcIInpTransfTherm}
\usage{
calcIInpTransfTherm()
}
\value{
OPENPROM input data iInpTransfTherm
}
\description{
Use ENERDATA input in electricity power plants data to derive OPENPROM input
parameter iInpTransfTherm
}
\examples{
\dontrun{
a <- calcOutput(type = "IInpTransfTherm", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
