% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIMatFacPlaAvailCap.R
\name{calcIMatFacPlaAvailCap}
\alias{calcIMatFacPlaAvailCap}
\title{calcIMatFacPlaAvailCap}
\usage{
calcIMatFacPlaAvailCap()
}
\value{
magpie object with OPENPROM input data iMatFacPlaAvailCap.
}
\description{
Use data from MENA_EDS model to derive OPENPROM input parameter iMatFacPlaAvailCap.
This dataset includes Maturty factors on Capacity.
}
\examples{
\dontrun{
a <- calcOutput(type = "IMatFacPlaAvailCap", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
