% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIMatrFactorData.R
\name{calcIMatrFactorData}
\alias{calcIMatrFactorData}
\title{calcIMatrFactorData}
\usage{
calcIMatrFactorData()
}
\value{
magpie object with OPENPROM input data iMatrFactorData.
}
\description{
Derive default values for iMatFacPlaAvailCap using uncalibrated data.
Provides maturity factors for all technologies across demand subsectors.
}
\examples{
\dontrun{
a <- calcOutput(type = "IMatrFactorData", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
