% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIMatureFacPlaDisp.R
\name{calcIMatureFacPlaDisp}
\alias{calcIMatureFacPlaDisp}
\title{calcIMatureFacPlaDisp}
\usage{
calcIMatureFacPlaDisp()
}
\value{
OPENPROM input data iIMatureFacPlaDisp.
}
\description{
Maturity factor related to plant dispatching to derive OPENPROM input
parameter iMatureFacPlaDisp
}
\examples{
\dontrun{
a <- calcOutput(type = "IMatureFacPlaDisp", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
