% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIMaxResPot.R
\name{calcIMaxResPot}
\alias{calcIMaxResPot}
\title{calcIMaxResPot}
\usage{
calcIMaxResPot()
}
\value{
OPENPROM input data iMaxResPot
The output data for EU countries calculated from the "EUROPEAN COMMISSION".
The output data for middle East and north Africa calculated from MENA_EDS
model. Countries with NA is used the region or global mean value.
}
\description{
Use res max potential data from the "EUROPEAN COMMISSION" and MENA_EDS model data,
to derive OPENPROM input parameter iMaxResPot.
}
\examples{
\dontrun{
a <- calcOutput(type = "IMaxResPot", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
