% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIMinResPot.R
\name{calcIMinResPot}
\alias{calcIMinResPot}
\title{calcIMinResPot}
\usage{
calcIMinResPot()
}
\value{
OPENPROM input data iMinResPot
The output data for EU countries calculated from the "EUROPEAN COMMISSION".
The output data for middle East and north Africa calculated from MENA_EDS
model. Countries with NA is used the region or global mean value.
}
\description{
The res min potential data calculated from res max potential data by
deviding the max by two, to derive OPENPROM input parameter iMinResPot.
The res max potential data is from the "EUROPEAN COMMISSION" and 
MENA_EDS model data.
}
\examples{
\dontrun{
a <- calcOutput(type = "IMinResPot", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
