% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcINatGasPriProElst.R
\name{calcINatGasPriProElst}
\alias{calcINatGasPriProElst}
\title{calcINatGasPriProElst}
\usage{
calcINatGasPriProElst()
}
\value{
OPENPROM input data iNatGasPriProElst.
}
\description{
Rate of coverage of natural gas consumption by domestic production, to derive
OPENPROM input parameter iNatGasPriProElst.
The data is from the "Enerdata".
}
\examples{
\dontrun{
a <- calcOutput(type = "INatGasPriProElst", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
