% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcINDISFuelCons.R
\name{calcINDISFuelConsumptionIEA}
\alias{calcINDISFuelConsumptionIEA}
\title{calcINDISFuelConsumptionIEA}
\usage{
calcINDISFuelConsumptionIEA(convfact = 1)
}
\value{
magpie object with OPENPROM input data iDataConsEneBranch.
}
\description{
Calculates fuel consumption in the Iron and Steel sector, corresponding to the 
variable "iFuelConsINDSE" for the "IS" sector in Open PROM. The calculation is 
performed for key regions according to two IEA scenarios: Stated Policies (IEA STEPS) 
and Sustainable Development (IEA SDS).
}
\details{
The function provides:
- A common value for fuel consumption in 2019 across both scenarios.
- Separate values for 2050 under IEA STEPS and IEA SDS.

The calculation is performed for each fuel type by determining the consumption 
for each technological route (e.g., BF-BOF, DR-EAF, etc.), using:
1. Technological shares (%) per region from the IEA Iron and Steel Roadmap.
2. Steel production from WEO 2023 Extended Data in Mton of steel per year
3. Specific fuel consumption per technological route from IEA The Future of Hydrogen in PJ/Mton of steel
The results of fuel consumption for fuel type in PJ/year is converted in Mtoe/year

Current Limitations:
- The function currently includes coal, natural gas, and electricity consumption, 
  as fuel-specific consumption values are globally defined without regional differentiation.
- Since the final consumption of fuel is based on generic fuel specific consumption per technological 
  route, the consumption for 2019 can be slightly underestimate, especially in country with old BF-BOF plants
- Future improvements could extend the calculation to oil, biomass, and heat using 
  country-level fuel shares from the IEA Iron and Steel Roadmap. It can be introduced country-specific parameters
  for calibrating the consumption of 2019 according to the energy balances
- Coal consumption includes usage in blast furnaces (BF), which is currently 
  classified under transformations in Open PROM.
- Further regional disaggregation may be necessary for MEA, Africa, and Central 
  and South America. However, for Europe, PRIMES data is available, and additional 
  disaggregation may introduce significant approximations.
}
\examples{
\dontrun{
a <- calcOutput(type = "INDISFuelConsumptionIEA", aggregate = FALSE)
}
}
\author{
Sonja Sechi, Fotis Sioutas
}
