% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIO.R
\name{calcIO}
\alias{calcIO}
\title{calcIO}
\usage{
calcIO(subtype = c("output", "output_biomass"))
}
\arguments{
\item{subtype}{Data subtype.}
}
\value{
IEA data as MAgPIE object aggregated to country level
}
\description{
Computes IEA-based model data for "output" and "output_biomass" "subtypes" by use of raw IEA "Energy Balances" data
and a mapping that corresponds to the structure of "products" and "flows" of IEA.
}
\details{
Mapping structure example: IEA product ANTCOAL used for IEA flow TPATFUEL, contributes via REMIND technology
coaltr for generating sesofos from pecoal (REMIND names)
}
\examples{
\dontrun{
a <- calcOutput("IO", subtype = "output")
}

}
\author{
Anastasis Giannousakis
}
