% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIPriceFuelsIntBase.R
\name{calcIPriceFuelsIntBase}
\alias{calcIPriceFuelsIntBase}
\title{calcIPriceFuelsIntBase}
\usage{
calcIPriceFuelsIntBase()
}
\value{
magpie object with OPENPROM input data iPriceFuelsIntBase.
}
\description{
Unlike iPriceFuelsInt, which includes several fossil fuel categories,
iPriceFuelsIntBase retains only the international price series for crude
oil (WCRO) and natural gas (WNGS). These fuels are selected from the
complete fuel price dataset generated by calcOutput(type = "IPriceFuelsInt",
aggregate = FALSE).
The resulting dataset provides the baseline international fuel price
trajectories for crude oil and natural gas used in OPEN-PROM, expressed
in $2015/toe and returned as a MAgPIE object.
}
\examples{
\dontrun{
a <- calcOutput(type = "IPriceFuelsIntBase", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
