% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calciResHeatCapFac.R
\name{calciResHeatCapFac}
\alias{calciResHeatCapFac}
\title{calciResHeatCapFac}
\usage{
calciResHeatCapFac()
}
\value{
magpie object with OPENPROM input data iResHeatCapFac
}
\description{
Calibrate average residential heating capacity factors for 2024 using EurostatHDD according to 2010 data from HotMaps
}
\examples{
\dontrun{
a <- calcOutput(type = "iResHeatCapFac", aggregate = FALSE)
}

}
\author{
Margarita Efthymiadou, Fotis Sioutas
}
