% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIShareTTechFuel.R
\name{calcIShareTTechFuel}
\alias{calcIShareTTechFuel}
\title{calcIShareTTechFuel}
\usage{
calcIShareTTechFuel()
}
\arguments{
\item{subtype}{string}
}
\value{
OPENPROM input data iFuelConsXXX
}
\description{
calcIShareTTechFuel
}
\examples{
\dontrun{
a <- calcOutput(type = "calcIShareTTechFuel", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
