% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISuppExports.R
\name{calcISuppExports}
\alias{calcISuppExports}
\title{calcISuppExports}
\usage{
calcISuppExports()
}
\value{
magpie object with OPENPROM input data iSuppExports.
}
\description{
Use data from ENERDATA to derive OPENPROM input parameter iSuppExports.
This dataset includes export values for each region and fuel type in Mtoe.
}
\examples{
\dontrun{
a <- calcOutput(type = "ISuppExports", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
