% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISuppPrimprod.R
\name{calcISuppPrimprod}
\alias{calcISuppPrimprod}
\title{calcISuppPrimprod}
\usage{
calcISuppPrimprod()
}
\value{
OPENPROM input data iSuppPrimProd.
}
\description{
Use ENERDATA Primary production data to derive OPENPROM input parameter iSuppPrimProd.
}
\examples{
\dontrun{
a <- calcOutput(type = "ISuppPrimprod", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
