% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISuppRatePrimProd.R
\name{calcISuppRatePrimProd}
\alias{calcISuppRatePrimProd}
\title{calcISuppRatePrimProd}
\usage{
calcISuppRatePrimProd()
}
\value{
OPENPROM input data iSuppRatePrimProd.
}
\description{
Use IEA data to derive OPENPROM input parameter iSuppRatePrimProd.
It is the ratio of primary production over the total energy supply
for each country.
}
\examples{
\dontrun{
a <- calcOutput(type = "ISuppRatePrimProd", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
