% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISuppRefCapacity.R
\name{calcISuppRefCapacity}
\alias{calcISuppRefCapacity}
\title{calcISuppRefCapacity}
\usage{
calcISuppRefCapacity()
}
\value{
magpie object with OPENPROM input data iSuppRefCapacity.
}
\description{
Use data from ENERDATA to derive OPENPROM input parameter iSuppRefCapacity
This dataset includes the refineries capacity in Mbl/day.
}
\examples{
\dontrun{
a <- calcOutput(type = "ISuppRefCapacity", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
