% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISupRateEneBranCons.R
\name{calcISupRateEneBranCons}
\alias{calcISupRateEneBranCons}
\title{calcISupRateEneBranCons}
\usage{
calcISupRateEneBranCons()
}
\value{
OPENPROM input data ISupRateEneBranCons.
}
\description{
Use ENERDATA Own Use of Energy Industries divided by Production data to 
derive the Rate of Energy Branch Consumption over total transformation output.
The output is the OPENPROM input parameter iSupRateEneBranCons.
}
\examples{
\dontrun{
a <- calcOutput(type = "ISupRateEneBranCons", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
