% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMACC.R
\name{calcMACC}
\alias{calcMACC}
\title{calcMACC}
\usage{
calcMACC()
}
\value{
A magpie object having baseline emissions and marginal abatement costs
for CH4, N2O and F-gases
}
\description{
Construct OPENPROM input data from the IMAGE MACC dataset.
}
\details{
This function reads the IMAGE Marginal Abatement Cost Curve (MACC) data
and converts baseline emissions and marginal abatement cost curves into
**MAgPIE**-compatible magpie objects. It processes: baseline emissions for all gases,
 MAC curves for CH\eqn{_4} and N\eqn{_2}O, absolute MAC curves for HFCs and PFCs,
 relative MAC curves for SF\eqn{_6}.

The function: 1) Reads the MACC data source using \code{readSource()}, 
converts baseline emissions and MAC curves to magpie format,
interpolates all time series to a common annual grid,
aggregates data from IMAGE regions to OPENPROM regions,
reduces MAC curve resolution using predefined or optimized carbon price points.

Helper functions are used to parse and normalize MAC curves and to
conservatively select representative cost points across all sectors.
}
\examples{
\dontrun{
a <- calcOutput(type = "MACC", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Alexandros Tsimpoukis
}
