% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPolicyValidationChecks.R
\name{calcPolicyValidationChecks}
\alias{calcPolicyValidationChecks}
\title{Calculate policy validation checks}
\usage{
calcPolicyValidationChecks()
}
\value{
A madrat result containing the checks as a data frame.
}
\description{
Reads the Climate Policy Modelling Protocol source through madrat and
returns the corresponding Postprom policy validation checks in memory.
Always use with aggregate=FALSE, since it is a dataframe.
}
\examples{
\dontrun{
a <- calcOutput("PolicyValidationChecks",aggregate = FALSE)
write.csv(a,"policy-validation-checks.csv",row.names = FALSE,na = "",quote = TRUE)
}

}
\author{
Alexandros Tsimpoukis
}
