% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPOP.R
\name{calcPOP}
\alias{calcPOP}
\title{calcPOP}
\usage{
calcPOP(scenario = "SSP2")
}
\arguments{
\item{scenario}{SSP scenario choice
Available scenario are:
\itemize{
\item `SSP1`:
\item `SSP2`:
\item `SSP3`:
\item `SSP4`:
\item `SSP5`:
}}
}
\value{
The SSP data filtered by pop
}
\description{
Derive population data based on SSP scenarios.
The data is sourced externally via calcOutput("Population") from the mrdrivers library,
and converted from millions to billions, with missing years interpolated to provide a
complete time series from 2010 to 2100.
}
\examples{
\dontrun{
POP <- calcOutput("POP", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
