% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrimaryEnergyPrice.R
\name{calcPrimaryEnergyPrice}
\alias{calcPrimaryEnergyPrice}
\title{calcPrimaryEnergyPrice}
\usage{
calcPrimaryEnergyPrice()
}
\value{
OPENPROM input data PrimaryEnergyPrice
}
\description{
calcPrimaryEnergyPrice
}
\examples{
\dontrun{
a <- calcOutput(type = "PrimaryEnergyPrice")
}

}
\author{
Fotis Sioutas
}
