% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrimes.R
\name{calcPrimes}
\alias{calcPrimes}
\title{calcPrimes}
\usage{
calcPrimes()
}
\value{
Primes fuel consumption in DOMSE, INDSE, NENSE, TRANSE
}
\description{
Derive fuel consumption data from PRIMES for the DOMSE, INDSE, NENSE, and TRANSE sectors.
The function combines PRIMES transport energy demand and PRIMES energy balance data,
maps fuels and sectors to the OPEN-PROM classification, and harmonises the data across
countries, fuels, and time. Missing values are interpolated and completed to ensure a
full time series (2010–2100).
}
\examples{
\dontrun{
a <- calcOutput(type = "Primes", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
