% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrimesPG.R
\name{calcPrimesPG}
\alias{calcPrimesPG}
\title{calcPrimesPG}
\usage{
calcPrimesPG(subtype = "SE")
}
\arguments{
\item{subtype}{"SE" for SE Electricity or "power generation" for
Gross Electricity generation by plant type and "capacity" for
Net Installed Power Capacity}
}
\value{
Primes PG data
}
\description{
Use Primes data Secondary Energy Electricity, Gross Electricity generation by plant type
and Net Installed Power Capacity
}
\examples{
\dontrun{
a <- calcOutput(type = "PrimesPG",subtype = "SE", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
