% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrimesPrices.R
\name{calcPrimesPrices}
\alias{calcPrimesPrices}
\title{calcPrimesPrices}
\usage{
calcPrimesPrices()
}
\value{
Primes Prices data for END USER PRICE
}
\description{
Derive end-user energy prices from the PRIMES model.
The function maps PRIMES price categories and sectors to the OPEN-PROM
classification, aggregates across fuels and sectors, and interpolates missing
values to ensure a continuous time series (up to 2070). Data are reported in
€2015 per toe and harmonised across regions.
}
\examples{
\dontrun{
a <- calcOutput(type = "PrimesPrices", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
