% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProdHistIS.R
\name{calcProdHistIS}
\alias{calcProdHistIS}
\title{calcProdHistIS}
\usage{
calcProdHistIS()
}
\value{
The read-in data into a magpie object.
}
\description{
Derive USGS_RAW_STEEL in Mt
}
\examples{
\dontrun{
a <- calcOutput(type = "ProdHistIS", aggregate = TRUE)
}
}
\author{
Fotis Sioutas
}
