% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSecEnElec.R
\name{calcSecEnElec}
\alias{calcSecEnElec}
\title{calcSecEnElec}
\usage{
calcSecEnElec()
}
\value{
Secondary Energy Electricity
}
\description{
Use Primes and Navigate data for Secondary Energy Electricity
}
\examples{
\dontrun{
a <- calcOutput(type = "SecEnElec", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
