% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSharesFuelPrices.R
\name{calcSharesFuelPrices}
\alias{calcSharesFuelPrices}
\title{calcSharesFuelPrices}
\usage{
calcSharesFuelPrices()
}
\value{
SharesFuelPrices
}
\description{
calcSharesFuelPrices
}
\examples{
\dontrun{
SharesFuelPrices <- calcOutput("SharesFuelPrices", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
