% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTFuelCons.R
\name{calcTFuelCons}
\alias{calcTFuelCons}
\title{calcTFuelCons}
\usage{
calcTFuelCons()
}
\value{
OPENPROM input targets iFuelCons
}
\description{
Fuel consumption share projections
Aggregates projected fuel consumption shares by fuel type from multiple
sectoral projection sources (domestic, industrial and NEN,PCHsectors) to the
country (ISO3) level. Aggregation is performed using fuel-consumption-based
weights derived from historical fuel consumption data.
}
\examples{
\dontrun{
a <- calcOutput(type = "TFuelCons")
}

}
\author{
Fotis Sioutas
}
