% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTFuelConsShares.R
\name{calcTFuelConsShares}
\alias{calcTFuelConsShares}
\title{calcTFuelConsShares}
\usage{
calcTFuelConsShares(subtype = "TRANSE")
}
\value{
OPENPROM input data iFuelConsXXX
}
\description{
calcTFuelConsShares
}
\examples{
\dontrun{
a <- calcOutput(type = "TFuelConsShares")
}

}
\author{
Fotis Sioutas
}
