% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransportCosts.R
\name{calcTransportCosts}
\alias{calcTransportCosts}
\title{calcTransportCosts}
\usage{
calcTransportCosts()
}
\value{
Read TechCosts data and convert it to a csv file
}
\description{
calcTransportCosts
}
\examples{
\dontrun{
a <- calcOutput(type = "TransportCosts", file = "iTransportCosts.csv", aggregate = FALSE)
}
}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
