% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTTransport.R
\name{calcTTansport}
\alias{calcTTansport}
\title{calcTTransport}
\usage{
calcTTansport()
}
\description{
Use Primes for
}
\examples{
\dontrun{
a <- calcOutput(type = "TTransport", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
