% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTTransport.R
\name{calcTTransport}
\alias{calcTTransport}
\title{calcTTransport}
\usage{
calcTTransport()
}
\value{
Magpie object for targets Transport
}
\description{
Use PrimesNewTransport and IEA_EV to calculate the OPENPROM input targets Transport
}
\examples{
\dontrun{
a <- calcOutput(type = "TTransport", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
