% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUN.R
\name{calcUN}
\alias{calcUN}
\title{calcUN}
\usage{
calcUN()
}
\value{
UN fuel consumption in DOMSE, INDSE, NENSE, TRANSE
}
\description{
Use UN data for fuel consumption in DOMSE, INDSE, NENSE, TRANSE
}
\examples{
\dontrun{
a <- calcOutput(type = "UN", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
