% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUSGS_RAW_STEEL.R
\name{calcUSGS_RAW_STEEL}
\alias{calcUSGS_RAW_STEEL}
\title{calcUSGS_RAW_STEEL}
\usage{
calcUSGS_RAW_STEEL()
}
\value{
The read-in data into a magpie object.
}
\description{
Derive USGS_RAW_STEEL in Mt
}
\examples{
\dontrun{
a <- calcOutput(type = "USGS_RAW_STEEL", aggregate = TRUE)
}
}
\author{
Fotis Sioutas
}
