% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertClimateWatch.R
\name{convertClimateWatch}
\alias{convertClimateWatch}
\title{convertClimateWatch}
\usage{
convertClimateWatch(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "ClimateWatch" data with spatial entries for each country.
}
\description{
The ISO codes of "ClimateWatch" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("ClimateWatch", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Alexandros Tsimpoukis
}
