% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEDGAR.R
\name{convertEDGAR}
\alias{convertEDGAR}
\title{convertEDGAR}
\usage{
convertEDGAR(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "EDGAR" data with spatial entries for each country.
}
\description{
The ISO codes of "EDGAR" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("EDGAR", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
