% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEDGAR2.R
\name{convertEDGAR2}
\alias{convertEDGAR2}
\title{convertEDGAR2}
\usage{
convertEDGAR2(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "EDGAR2" data with spatial entries for each country.
}
\description{
The ISO codes of "EDGAR2" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("EDGAR2", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
