% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEMBER.R
\name{convertEMBER}
\alias{convertEMBER}
\title{convertEMBER}
\usage{
convertEMBER(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "EMBER" data with spatial entries for each country.
}
\description{
The ISO codes of "EMBER" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("EMBER", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
