% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertENGAGE.R
\name{convertENGAGE}
\alias{convertENGAGE}
\title{convertENGAGE}
\usage{
convertENGAGE(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read in data from the ENGAGE project.
The dataset contains carbon price data for the current policies scenario GP_CurPol_T45.
This function uses toolCountryFill to fill missing country data with 0 and
assign to all EU28 countries the value of EUR.
}
\examples{
\dontrun{
a <- readSource("ENGAGE", convert = TRUE)
}


}
\author{
Anastasis Giannousakis
}
