% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEU_COM_RES.R
\name{convertEU_COM_RES}
\alias{convertEU_COM_RES}
\title{convertEU_COM_RES}
\usage{
convertEU_COM_RES(x)
}
\arguments{
\item{x}{MAgPIE object with ISO country codes.}
}
\value{
The "EU_COM_RES" data with spatial entries for each country.
}
\description{
The ISO codes of "EU_COM_RES" data are compared with the official ISO
code country list by using the function toolCountryFill of madrat.
}
\examples{
\dontrun{
a <- readSource("EU_COM_RES", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
