% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEU_RefScen2020.R
\name{convertEU_RefScen2020}
\alias{convertEU_RefScen2020}
\title{convertEU_RefScen2020}
\usage{
convertEU_RefScen2020(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read in data from the EU_RefScen2020.
The dataset contains carbon price data for the EU Reference Scenario 2020.
This function uses toolCountryFill to fill missing country data with 0 and
assign to all EU28 countries the value of EUR.
}
\examples{
\dontrun{
a <- readSource("EU_RefScen2020", convert = TRUE)
}


}
\author{
Anastasis Giannousakis
}
