% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEURefPlus.R
\name{convertEURefPlus}
\alias{convertEURefPlus}
\title{convertEURefPlus}
\usage{
convertEURefPlus(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "EURefPlus" data with spatial entries for each country.
}
\description{
The ISO codes of "EURefPlus" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("EURefPlus", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
