% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEuropeanHydrogen.R
\name{convertEuropeanHydrogen}
\alias{convertEuropeanHydrogen}
\title{convertEuropeanHydrogen}
\usage{
convertEuropeanHydrogen(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "European Hydrogen Observatory" data with spatial entries for each country.
}
\description{
The ISO codes of "EuropeanHydrogen" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("EuropeanHydrogen", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
