% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEuropeanHydrogenPrices.R
\name{convertEuropeanHydrogenPrices}
\alias{convertEuropeanHydrogenPrices}
\title{convertEuropeanHydrogenPrices}
\usage{
convertEuropeanHydrogenPrices(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "European Hydrogen Observatory" data with spatial entries for each country.
}
\description{
The ISO codes of "EuropeanHydrogenPrices" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("EuropeanHydrogenPrices", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
