% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurostat2.R
\name{convertEurostat2}
\alias{convertEurostat2}
\title{convertEurostat2}
\usage{
convertEurostat2(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "EUROSTAT" data with spatial entries for each country.
}
\description{
convertEurostat2
}
\examples{
\dontrun{
a <- readSource("Eurostat2", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
