% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurostat_ELVS.R
\name{convertEurostat_ELVS}
\alias{convertEurostat_ELVS}
\title{convertEurostat_ELVS}
\usage{
convertEurostat_ELVS(x)
}
\arguments{
\item{x}{MAgPIE object with ISO country codes.}
}
\value{
The "Eurostat_ELVS" data with spatial entries for each country.
}
\description{
The ISO codes of the data are compared with the official ISO
code country list by using the function toolCountryFill of madrat.
}
\examples{
\dontrun{
a <- readSource("Eurostat_ELVS", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
