% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurostatHDD.R
\name{convertEurostatHDD}
\alias{convertEurostatHDD}
\title{convertEurostatHDD}
\usage{
convertEurostatHDD(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "EurostatHDD" data with spatial entries for each country.
}
\description{
The ISO codes of "EurostatHDD" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("EurostatHDD", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
