% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurostatHP.R
\name{convertEurostatHP}
\alias{convertEurostatHP}
\title{convertEurostatHP}
\usage{
convertEurostatHP(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "EurostatHP" data with spatial entries for each country.
}
\description{
The ISO codes of "EurostatHP" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("EurostatHP", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
