% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAO.R
\name{convertFAO}
\alias{convertFAO}
\title{convertFAO}
\usage{
convertFAO(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "FAO" data with spatial entries for each country.
}
\description{
The ISO codes of "FAO" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("FAO", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
