% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAOFertilizers.R
\name{convertFAOFertilizers}
\alias{convertFAOFertilizers}
\title{convertFAOFertilizers}
\usage{
convertFAOFertilizers(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "FAO" data with spatial entries for each country.
}
\description{
The ISO codes of "FAOFertilizers" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("FAOFertilizers", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
