% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAOFishing.R
\name{convertFAOFishing}
\alias{convertFAOFishing}
\title{convertFAOFishing}
\usage{
convertFAOFishing(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "FAO" data with spatial entries for each country.
}
\description{
The ISO codes of "FAOFishing" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("FAOFishing", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
