% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAOForestry.R
\name{convertFAOForestry}
\alias{convertFAOForestry}
\title{convertFAOForestry}
\usage{
convertFAOForestry(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "FAOForestry" data with spatial entries for each country.
}
\description{
The ISO codes of "FAOForestry" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("FAOForestry", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
