% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAOLandUse.R
\name{convertFAOLandUse}
\alias{convertFAOLandUse}
\title{convertFAOLandUse}
\usage{
convertFAOLandUse(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "FAO" data with spatial entries for each country.
}
\description{
The ISO codes of "FAOLandUse" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("FAOLandUse", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
