% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAOProductionCrops.R
\name{convertFAOProductionCrops}
\alias{convertFAOProductionCrops}
\title{convertFAOProductionCrops}
\usage{
convertFAOProductionCrops(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "FAOProductionCrops" data with spatial entries for each country.
}
\description{
The ISO codes of "FAOProductionCrops" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("FAOProductionCrops", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
