% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAOprojections.R
\name{convertFAOprojections}
\alias{convertFAOprojections}
\title{convertFAOprojections}
\usage{
convertFAOprojections(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "FAO" data with spatial entries for each country.
}
\description{
The ISO codes of "FAO projections" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("FAOprojections", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
