% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGEM.R
\name{convertGEM}
\alias{convertGEM}
\title{convertGEM}
\usage{
convertGEM(x)
}
\arguments{
\item{x}{Quitte object with ISO country codes.}
}
\value{
The "GEM" data with spatial entries for each country.
}
\description{
The ISO codes of "GEM" data are compared with the official ISO
code country list.
}
\examples{
\dontrun{
a <- readSource("GEM", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
