% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGEME3.R
\name{convertGEME3}
\alias{convertGEME3}
\title{convertGEME3}
\usage{
convertGEME3(x)
}
\arguments{
\item{x}{MAgPIE object with ISO country codes.}
}
\value{
The "GEME3" data with spatial entries for each country.
}
\description{
The ISO codes of "GEME3" data are compared with the official ISO
code country list by using the function toolCountryFill of madrat.
}
\examples{
\dontrun{
a <- readSource("GEME3", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
