% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGidden.R
\name{convertGidden}
\alias{convertGidden}
\title{convertGidden}
\usage{
convertGidden(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "Gidden" data with spatial entries for each country.
}
\description{
The ISO codes of "Gidden" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("Gidden", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
