% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGLOBIOMEU.R
\name{convertGLOBIOMEU}
\alias{convertGLOBIOMEU}
\title{convertGLOBIOMEU}
\usage{
convertGLOBIOMEU(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "GLOBIOMEU" data with spatial entries for each country.
}
\description{
The ISO codes of "GLOBIOMEU" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("GLOBIOMEU", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
