% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHotMaps.R
\name{convertHotMaps}
\alias{convertHotMaps}
\title{convertEurostatHDD}
\usage{
convertHotMaps(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "EurostatHDD" data with spatial entries for each country.
}
\description{
The ISO codes of "EurostatHDD" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("HotMaps", convert = TRUE)
}

}
\author{
Margarita Efthymiadou
}
