% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertICTPrisma.R
\name{convertICTPrisma}
\alias{convertICTPrisma}
\title{convertICTPrisma}
\usage{
convertICTPrisma(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "ICTPrisma" data with spatial entries for each country.
}
\description{
The ISO codes data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("ICTPrisma", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
