% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA.R
\name{convertIEA}
\alias{convertIEA}
\title{convertIEA}
\usage{
convertIEA(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEA" data with spatial entries for each country.
}
\description{
The ISO codes of "IEA" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("IEA", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
