% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA2024.R
\name{convertIEA2024}
\alias{convertIEA2024}
\title{convertIEA2024}
\usage{
convertIEA2024(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEA2024" data with spatial entries for each country.
}
\description{
The ISO codes of "IEA2024" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("IEA2024", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
