% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA2025.R
\name{convertIEA2025}
\alias{convertIEA2025}
\title{convertIEA2025}
\usage{
convertIEA2025(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEA2024" data with spatial entries for each country.
}
\description{
The ISO codes of "IEA2025" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("IEA2025", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
