% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_CO2.R
\name{convertIEA_CO2}
\alias{convertIEA_CO2}
\title{convertIEA_CO2}
\usage{
convertIEA_CO2(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEA_CO2" data with spatial entries for each country.
}
\description{
The ISO codes of "IEA_CO2" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("IEA_CO2", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
