% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_Energy_Projections_Balances.R
\name{convertIEA_Energy_Projections_Balances}
\alias{convertIEA_Energy_Projections_Balances}
\title{convertIEA_Energy_Projections_Balances}
\usage{
convertIEA_Energy_Projections_Balances(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEA_Energy_Projections_Balances" data with spatial entries for each country.
}
\description{
The ISO codes of "IEA_Energy_Projections_Balances" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("IEA_Energy_Projections_Balances", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
