% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/convertIEA_Energy_Projections_Extended_Indicators.R
\name{convertIEA_Energy_Projections_Extended_Indicators}
\alias{convertIEA_Energy_Projections_Extended_Indicators}
\title{convertIEA_Energy_Projections_Extended_Indicators}
\usage{
convertIEA_Energy_Projections_Extended_Indicators(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEA_Energy_Projections_Extended_Indicators" data with spatial entries for each country.
}
\description{
The ISO codes of "IEA_Energy_Projections_Extended_Indicators" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("IEA_Energy_Projections_Extended_Indicators", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
