% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_EV.R
\name{convertIEA_EV}
\alias{convertIEA_EV}
\title{convertIEA_EV}
\usage{
convertIEA_EV(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEA_EV" data with spatial entries for each country.
}
\description{
The ISO codes of "IEA_EV" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("IEA_EV", convert = TRUE)
}

}
\author{
Fotis Sioutas, Michael Madianos
}
