% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_WEO_2025_ExtendedData.R
\name{convertIEA_WEO_2025_ExtendedData}
\alias{convertIEA_WEO_2025_ExtendedData}
\title{convertIEA_WEO_2025_ExtendedData}
\usage{
convertIEA_WEO_2025_ExtendedData(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEA_WEO_2025_ExtendedData" data with spatial entries for each country.
}
\description{
convertIEA_WEO_2025_ExtendedData
}
\examples{
\dontrun{
a <- readSource("IEA_WEO_2025_ExtendedData", subtype = "IEA_WEO_2025_ExtendedData", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
