% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_WEO_TechCosts.R
\name{convertIEA_WEO_TechCosts}
\alias{convertIEA_WEO_TechCosts}
\title{convertIEA_WEO_TechCosts}
\usage{
convertIEA_WEO_TechCosts(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEA_WEO_TechCosts" data with spatial entries for each country.
}
\description{
The ISO codes of "IEA_WEO_TechCosts" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("IEA_WEO_TechCosts", subtype = "Renewables", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
