% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEACrudeImportCountry.R
\name{convertIEACrudeImportCountry}
\alias{convertIEACrudeImportCountry}
\title{convertIEACrudeImportCountry}
\usage{
convertIEACrudeImportCountry(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEACrudeImportCountry" data with spatial entries for each country.
}
\description{
The ISO codes of "IEACrudeImportCountry" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("IEACrudeImportCountry", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
