% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEAEnergyPrices.R
\name{convertIEAEnergyPrices}
\alias{convertIEAEnergyPrices}
\title{convertIEAEnergyPrices}
\usage{
convertIEAEnergyPrices(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEAEnergyPrices" data with spatial entries for each country.
}
\description{
The ISO codes of "IEAEnergyPrices" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("IEAEnergyPrices", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
