% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEATOTPRICES.R
\name{convertIEATOTPRICES}
\alias{convertIEATOTPRICES}
\title{convertIEATOTPRICES}
\usage{
convertIEATOTPRICES(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IEAEnergyPrices" data with spatial entries for each country.
}
\description{
The ISO codes of "IEATOTPRICES" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("IEATOTPRICES", convert = TRUE)
}

}
\author{
Michael Madianos, Fotis Sioutas
}
