% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIMF.R
\name{convertIMF}
\alias{convertIMF}
\title{convertIMF}
\usage{
convertIMF(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IMF" data with spatial entries for each country.
}
\description{
The ISO codes of "GDP IMF" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("IMF", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
