% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIRF.R
\name{convertIRF}
\alias{convertIRF}
\title{convertIRF}
\usage{
convertIRF(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IRF" data with spatial entries for each country.
}
\description{
Add missing periods to data frame and interpolate missing values.
The ISO codes of "IRF" data are compared with the official ISO code country list.
NA values are replaced with zeros and historical data are corrected.
}
\examples{
\dontrun{
a <- readSource("IRF", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
