% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertiTEM.R
\name{convertiTEM}
\alias{convertiTEM}
\title{convertiTEM}
\usage{
convertiTEM(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "iTEM" data with spatial entries for each country.
}
\description{
The ISO codes of "iTEM" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("iTEM", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
