% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNavigate.R
\name{convertNavigate}
\alias{convertNavigate}
\title{convertNavigate}
\usage{
convertNavigate(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "Navigate" data with spatial entries for each country.
}
\description{
The ISO codes of "Navigate" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("Navigate", subtype = "SUP_NPi_Default", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
