% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNavigateConsumption.R
\name{convertNavigateConsumption}
\alias{convertNavigateConsumption}
\title{convertNavigateConsumption}
\usage{
convertNavigateConsumption(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "Navigate" data with spatial entries for each country.
}
\description{
The ISO codes of "Navigate" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("NavigateConsumption", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
