% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertOPEN_GEM.R
\name{convertOPEN_GEM}
\alias{convertOPEN_GEM}
\title{convertOPEN_GEM}
\usage{
convertOPEN_GEM(x)
}
\arguments{
\item{x}{MAgPIE object with ISO country codes.}
}
\value{
The "GEME3" data with spatial entries for each country.
}
\description{
The ISO codes of "GEME3" data are compared with the official ISO
code country list by using the function toolCountryFill of madrat.
}
\examples{
\dontrun{
a <- readSource("OPEN_GEM")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
