% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPIK.R
\name{convertPIK}
\alias{convertPIK}
\title{convertPIK}
\usage{
convertPIK(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "PIK" data with spatial entries for each country.
}
\description{
The ISO codes of "PIK" data are compared with the official ISO code country list.
NA values are replaced with zeros
}
\examples{
\dontrun{
a <- readSource("PIK", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
