% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUN.R
\name{convertUN}
\alias{convertUN}
\title{convertUN}
\usage{
convertUN(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "UN" data with spatial entries for each country.
}
\description{
The ISO codes of "UN" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("UN", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
