% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUSGS.R
\name{convertUSGS}
\alias{convertUSGS}
\title{convertUSGS}
\usage{
convertUSGS(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "USGS" data with spatial entries for each country.
}
\description{
The ISO codes of "USGS" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("USGS", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
