% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWDI_PA.R
\name{convertWDI_PA}
\alias{convertWDI_PA}
\title{convertWDI_PA}
\usage{
convertWDI_PA(x)
}
\arguments{
\item{x}{MAgPIE object with ISO country codes.}
}
\value{
The "WDI_PA" data with spatial entries for each country.
}
\description{
The ISO codes of "WDI_PA" data are compared with the official ISO
code country list by using the function toolCountryFill of madrat.
}
\examples{
\dontrun{
a <- readSource("WDI_PA", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
