% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWorldBankCarPr.R
\name{convertWorldBankCarPr}
\alias{convertWorldBankCarPr}
\title{convertWorldBankCarPr}
\usage{
convertWorldBankCarPr(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IRF" data with spatial entries for each country.
}
\description{
Add missing periods to data frame and interpolate missing values.
The ISO codes of "WorldBankCarPr" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("WorldBankCarPr", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
