% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWorldBankCarPr2025.R
\name{convertWorldBankCarPr2025}
\alias{convertWorldBankCarPr2025}
\title{convertWorldBankCarPr2025}
\usage{
convertWorldBankCarPr2025(x)
}
\arguments{
\item{x}{MAgPIE object.}
}
\value{
The "IRF" data with spatial entries for each country.
}
\description{
Add missing periods to data frame and interpolate missing values.
The ISO codes of "WorldBankCarPr" data are compared with the official ISO code country list.
}
\examples{
\dontrun{
a <- readSource("WorldBankCarPr2025", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
