% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptionmrprom.R
\name{descriptionmrprom}
\alias{descriptionmrprom}
\title{descriptionmrprom}
\usage{
descriptionmrprom(file)
}
\arguments{
\item{file}{The input.gms file path.}
}
\value{
The description of mrprom functions
}
\description{
Description and in which functions are used the functions of mrprom.
}
\examples{
\dontrun{
a <- descriptionmrprom(file)
}

}
\author{
Fotis Sioutas
}
