% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eurostat.R
\name{Eurostat}
\alias{Eurostat}
\title{Eurostat}
\usage{
Eurostat(subtypes = "nama_10_pe")
}
\arguments{
\item{subtypes}{Type of data.
Available types are:
\itemize{
\item `naio_10_cp1700`:Symmetric input-output table at basic prices (product by product)
\item `naio_10_cp15`:Supply table at basic prices incl. transformation into purchasers' prices
\item `naio_10_cp16`:Use table at purchasers' prices
\item `naio_10_cp1620`:Table of trade and transport margins
\item `naio_10_cp1630`:Table of taxes less subsidies on product
\item `bop_its6_det`:International trade in services
\item `nama_10_pe`:Population and employment
\item `nama_10_a64_e`:National accounts aggregates by industry
\item `demo_pjan`:Population on 1 January by age and sex
\item `lfsa_pganws`:Population by sex, age, citizenship and labour status
\item `lfsa_eisn2`:Employment by sex, age, occupation and economic activity
\item `lfsa_ugpis`:Previous occupations of the unemployed, by sex
\item `nasa_10_nf_tr`:Non-financial transactions - annual data
\item `nasa_10_f_tr`:Financial transactions - annual data
\item `env_ac_ainah_r2`:Air emissions accounts by NACE Rev. 2 activity
\item `nama_10_co3_p3`:Final consumption expenditure of households by consumption purpose
}}
}
\value{
The read-in data into a gdx file
}
\description{
Eurostat
}
\examples{
\dontrun{
a <- Eurostat(subtypes = c("nama_10_pe","naio_10_cp1630"))
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
