% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_historical.R
\name{full_historircal}
\alias{full_historircal}
\title{full_historircal}
\usage{
full_historircal()
}
\value{
The .mif objects
}
\description{
Creates .mif objects from calcoutput
}
\examples{
\dontrun{
a <- full_historircal()
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
