% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullHISTORICAL.R
\name{fullHISTORICAL}
\alias{fullHISTORICAL}
\title{fullHISTORICAL}
\usage{
fullHISTORICAL()
}
\value{
The .mif objects
}
\description{
Creates .mif objects from calcoutput
}
\examples{
\dontrun{
a <- retrieveData("Historical", regionmapping = "regionmappingOPDEV2.csv")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
