% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullOPEN-PROM.R
\name{fullOPEN_PROM}
\alias{fullOPEN_PROM}
\title{fullOPEN-PROM}
\usage{
fullOPEN_PROM()
}
\value{
The read-in data into a magpie object.
}
\description{
Read in several files with data from GEME3, IRF and WDI_PA and convert it
to a csv file.
The dataset contains several data types about transport, traffic, air
transport passengers per country and per year and Production Level and
Unit Cost data.
}
\examples{
\dontrun{
 a <- retrieveData("OPEN_PROM", regionmapping = "regionmappingOP.csv")
}
}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
