% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullTARGETS.R
\name{fullTARGETS}
\alias{fullTARGETS}
\title{fullTARGETS}
\usage{
fullTARGETS()
}
\value{
The read-in target data into a magpie object.
}
\description{
fullTARGETS
}
\examples{
\dontrun{
a <- retrieveData("TARGETS", regionmapping = "regionmappingOP.csv")
}
}
\author{
Michael Madianos, Anastasis Giannousakis
}
