% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullVALIDATION2.R
\name{fullVALIDATION2}
\alias{fullVALIDATION2}
\title{fullVALIDATION2}
\usage{
fullVALIDATION2()
}
\value{
The mif file
}
\description{
Read data from PRIMES and EMBER convert it to a mrprom mif
file so to compare output mif file with OPEN-PROM output.
}
\examples{
\dontrun{
a <- retrieveData("fullVALIDATION2", regionmapping = "regionmappingOPDEV3.csv")
}

}
\author{
Fotis Sioutas
}
