% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEA.R
\name{IEA}
\alias{IEA}
\title{IEA}
\usage{
IEA(subtype = "1960")
}
\arguments{
\item{subtype}{The year.}
}
\value{
The read-in data into a gdx file
}
\description{
IEA
}
\examples{
\dontrun{
a <- IEA(subtype = c("1960","1970"))
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
