% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_overview.R
\name{input_data_overview}
\alias{input_data_overview}
\title{input_data_overview}
\usage{
input_data_overview()
}
\value{
The description of mrprom functions
}
\description{
Description and in which functions are used the functions of mrprom.
}
\examples{
\dontrun{
a <- input_data_overview()
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
