% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingTask.R
\name{modellingTask}
\alias{modellingTask}
\title{modellingTask}
\usage{
modellingTask()
}
\value{
Info about sources of mrprom.
}
\description{
Modelling the sources of mrprom.
}
\examples{
\dontrun{
a <- modellingTask()
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
