% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readACHIEVE.R
\name{readACHIEVE}
\alias{readACHIEVE}
\title{readACHIEVE}
\usage{
readACHIEVE(subtype = "Cities")
}
\arguments{
\item{subtype}{Companies or Cities}
}
\value{
mif object with the requested output data
}
\description{
Read in a XLSX file from ACHIEVE and convert it to a magpie object.
}
\examples{
\dontrun{
a <- readSource("ACHIEVE", subtype = "Cities")
}

}
\author{
Fotis Sioutas, Dionysis
}
