% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAGENRES.R
\name{readAGENRES}
\alias{readAGENRES}
\title{readAGENRES}
\usage{
readAGENRES()
}
\value{
The read-in data into a magpie object.
}
\description{
Read AGENRES data energy use per activity.
}
\examples{
\dontrun{
a <- readSource("AGENRES")
}

}
\author{
Fotis Sioutas
}
