% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBoT.R
\name{readBoT}
\alias{readBoT}
\title{readBoT}
\usage{
readBoT()
}
\value{
magpie object with the requested output data about 
End-of-life passenger car vehicles.
}
\description{
Read in a csv file and convert it to a magpie object
The data has information about End-of-life passenger car vehicles.
}
\examples{
\dontrun{
a <- readSource("BoT")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
