% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCarPrSoCDRHighestAmbition.R
\name{readCarPrSoCDRHighestAmbition}
\alias{readCarPrSoCDRHighestAmbition}
\title{readCarPrSoCDRHighestAmbition}
\usage{
readCarPrSoCDRHighestAmbition()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read CarPrSoCDRHighestAmbition.
The dataset contains carbon price data.
}
\examples{
\dontrun{
a <- readSource("CarPrSoCDRHighestAmbition")
}

}
\author{
Fotis Sioutas
}
