% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClimateTrace.R
\name{readClimateTrace}
\alias{readClimateTrace}
\title{readClimateTrace}
\usage{
readClimateTrace()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Emissions|CO2|AFOLU of all world countries from ClimateTrace
}
\examples{
\dontrun{
a <- readSource("ClimateTrace", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
