% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClimateWatch.R
\name{readClimateWatch}
\alias{readClimateWatch}
\title{readClimateWatch}
\usage{
readClimateWatch(subtype = "historical_emissions_ClimateWatch")
}
\arguments{
\item{subtype}{Type of data that should be read. The type is referring to the
excel and converts it to a magpie object.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read ClimateWatch data : model meta information, scenario meta information
}
\examples{
\dontrun{
a <- readSource("ClimateWatch", subtype = "historical_emissions_ClimateWatch")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
