% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDGAGRIOUTLOOK.R
\name{readDGAGRIOUTLOOK}
\alias{readDGAGRIOUTLOOK}
\title{readDGAGRIOUTLOOK}
\usage{
readDGAGRIOUTLOOK()
}
\value{
magpie DGAGRIOUTLOOK factors.
}
\description{
Read in a XLSX file and convert it to a magpie.
The data has information about DGAGRIOUTLOOK factors.
}
\examples{
\dontrun{
a <- readSource("DGAGRIOUTLOOK")
}

}
\author{
Fotis Sioutas
}
