% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDieselGasTaxesEurope.R
\name{readDieselGasTaxesEurope}
\alias{readDieselGasTaxesEurope}
\title{readDieselGasTaxesEurope}
\usage{
readDieselGasTaxesEurope()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read in data Diesel Gas Taxes Europe.
The dataset contains Diesel Gas Taxes Europe data from taxfoundation.org/data/all/eu/diesel-gas-taxes-europe/.
}
\examples{
\dontrun{
a <- readSource("DieselGasTaxesEurope")
}

}
\author{
Fotis Sioutas
}
