% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readECEMF.R
\name{readECEMF}
\alias{readECEMF}
\title{readECEMF}
\usage{
readECEMF()
}
\value{
The read-in data into a magpie object.
}
\description{
Read ECEMF
}
\examples{
\dontrun{
a <- readSource("ECEMF")
}

}
\author{
Fotis Sioutas
}
