% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGAR.R
\name{readEDGAR}
\alias{readEDGAR}
\title{readEDGAR}
\usage{
readEDGAR()
}
\value{
The read-in data into a magpie object.
}
\description{
Read IEA EDGAR CO2 emissions from EDGAR.
}
\examples{
\dontrun{
a <- readSource("EDGAR", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
