% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGAR2.R
\name{readEDGAR2}
\alias{readEDGAR2}
\title{readEDGAR2}
\usage{
readEDGAR2()
}
\value{
The read-in data into a magpie object.
}
\description{
Read GHG emissions of all world countries, 2024 Report from EDGAR.
}
\examples{
\dontrun{
a <- readSource("EDGAR2", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
