% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEMBER.R
\name{readEMBER}
\alias{readEMBER}
\title{readEMBER}
\usage{
readEMBER()
}
\value{
The read-in data into a magpie object.
}
\description{
Read EMBER capacity and production.
}
\examples{
\dontrun{
a <- readSource("EMBER", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
