% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEMO.R
\name{readEMO}
\alias{readEMO}
\title{readEMO}
\usage{
readEMO()
}
\value{
The read-in data into a magpie object.
}
\description{
readEMO
}
\examples{
\dontrun{
a <- readSource("EMO")
}

}
\author{
Fotis Sioutas
}
