% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readENERDATA2.R
\name{readENERDATA2}
\alias{readENERDATA2}
\title{readENERDATA2}
\usage{
readENERDATA2(subtype = "electricity production")
}
\arguments{
\item{subtype}{string. By choosing a subtype you filter the ENERDATA dataset
(1800+ variables) by type, to allow further processing of specific variables}
}
\value{
The read-in data into a magpie object
}
\description{
Read a csv files with data from ENERDATA.
The dataset contains several energy-related data types from ENERDATA for
various countries and years.
}
\examples{
\dontrun{
a <- readSource("ENERDATA2", "electricity production")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
