% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readENGAGE.R
\name{readENGAGE}
\alias{readENGAGE}
\title{readENGAGE}
\usage{
readENGAGE()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read in data from the ENGAGE project.
The dataset contains carbon price data for the current policies scenario GP_CurPol_T45.
}
\examples{
\dontrun{
a <- readSource("ENGAGE")
}

}
\author{
Anastasis Giannousakis
}
