% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEU_COM_RES.R
\name{readEU_COM_RES}
\alias{readEU_COM_RES}
\title{readEU_COM_RES}
\usage{
readEU_COM_RES()
}
\value{
magpie object with the requested output data about
res potential of EU countries.
}
\description{
Read in a XLSX file and convert it to a magpie object
The data has information about res max potential of EU countries from the
EUROPEAN COMMISSION.
}
\examples{
\dontrun{
a <- readSource("EU_COM_RES")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
