% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEU_RefScen2020.R
\name{readEU_RefScen2020}
\alias{readEU_RefScen2020}
\title{readEU_RefScen2020}
\usage{
readEU_RefScen2020()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read in data from CarbonPrice_fromReportFig8.
The dataset contains carbon price data for the EU Reference Scenario 2020.
}
\examples{
\dontrun{
a <- readSource("EU_RefScen2020", convert = TRUE)
}

}
\author{
Anastasis Giannousakis
}
