% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEUC.R
\name{readEUC}
\alias{readEUC}
\title{readEUC}
\usage{
readEUC()
}
\value{
dataframe with the requested output data about emission factors from
the EUROPEAN COMMISSION.
}
\description{
Read in a XLSX file and convert it to a dataframe.
The data has information about emission factors from the EUROPEAN COMMISSION.
}
\examples{
\dontrun{
a <- readSource("EUC")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
