% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEURefPlus.R
\name{readEURefPlus}
\alias{readEURefPlus}
\title{readEURefPlus}
\usage{
readEURefPlus()
}
\value{
magpie object object with the requested output GDP|MER and Population
}
\description{
Read in an xls file and convert it to a magpie object
The data has information about GDP|MER and Population.
}
\examples{
\dontrun{
a <- readSource("EURefPlus")
}

}
\author{
Alexandros Tsimpoukis, Dionysis Pramangioulis
}
