% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEuropeanHydrogen.R
\name{readEuropeanHydrogen}
\alias{readEuropeanHydrogen}
\title{readEuropeanHydrogen}
\usage{
readEuropeanHydrogen()
}
\value{
The read-in data into a magpie object.
}
\description{
Read EuropeanHydrogen demand from European Hydrogen Observatory.
}
\examples{
\dontrun{
a <- readSource("EuropeanHydrogen", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
