% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEuropeanHydrogenPrices.R
\name{readEuropeanHydrogenPrices}
\alias{readEuropeanHydrogenPrices}
\title{readEuropeanHydrogenPrices}
\usage{
readEuropeanHydrogenPrices()
}
\value{
The read-in data into a magpie object.
}
\description{
Read EuropeanHydrogenPrices from European Hydrogen Observatory.
}
\examples{
\dontrun{
a <- readSource("EuropeanHydrogenPrices", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
