% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostat_ELVS.R
\name{readEurostat_ELVS}
\alias{readEurostat_ELVS}
\title{readEurostat_ELVS}
\usage{
readEurostat_ELVS()
}
\value{
magpie object with the requested output data about 
End-of-life vehicles.
}
\description{
Read in a csv file and convert it to a magpie object
The data has information about End-of-life vehicles.
}
\examples{
\dontrun{
a <- readSource("Eurostat_ELVS")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
