% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostat_Transport.R
\name{readEurostat_Transport}
\alias{readEurostat_Transport}
\title{readEurostat_Transport}
\usage{
readEurostat_Transport()
}
\value{
magpie object with the requested output data about transport passengers per country and per year
}
\description{
Read in a csv file and convert it to a magpie object
The data has information about transport passengers per country and per year
}
\examples{
\dontrun{
a <- readSource("Eurostat_Transport")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
