% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostatHDD.R
\name{readEurostatHDD}
\alias{readEurostatHDD}
\title{readEurostatHDD}
\usage{
readEurostatHDD()
}
\value{
The read-in data into a magpie object.
}
\description{
Read HDD from Eurostat
}
\examples{
\dontrun{
a <- readSource("EurostatHDD")
}

}
\author{
Fotis Sioutas
}
