% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostatHP.R
\name{readEurostatHP}
\alias{readEurostatHP}
\title{readEurostatHP}
\usage{
readEurostatHP()
}
\value{
The read-in data into a magpie object.
}
\description{
Read EurostatHP
}
\examples{
\dontrun{
a <- readSource("EurostatHP")
}

}
\author{
Fotis Sioutas
}
