% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEvalGlobal.R
\name{readEvalGlobal}
\alias{readEvalGlobal}
\title{readEvalGlobal}
\usage{
readEvalGlobal(file)
}
\arguments{
\item{file}{A gams file containing GAMS code.}
}
\value{
A vector of values the evalglobal variables are set to as names.
}
\description{
Reads all evalglobals given in a GAMS code and returns them.
}
\examples{
\dontrun{
readEvalGlobal <- function(main.gms)
}


}
\author{
Anastasis Giannousakis Fotis Sioutas
}
