% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAO.R
\name{readFAO}
\alias{readFAO}
\title{Read FAOSTAT Food Balance Sheet data}
\usage{
readFAO(subset = "Food supply (kcal/capita/day)")
}
\value{
A list containing the data as a magpie object and its metadata.
}
\description{
Reads the normalized FAOSTAT Food Balance Sheets CSV file and converts
it into a magpie object.
}
\examples{
\dontrun{
x <- readSource("FAO")
}

}
\author{
Fotis Sioutas
}
