% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOFertilizers.R
\name{readFAOFertilizers}
\alias{readFAOFertilizers}
\title{Read FAOSTAT FAOFertilizers}
\usage{
readFAOFertilizers()
}
\value{
A list containing the data as a magpie object and its metadata.
}
\description{
Reads the FAO FAOFertilizers CSV file and converts it into a magpie object.
}
\examples{
\dontrun{
x <- readSource("FAOFertilizers")
}

}
\author{
Fotis Sioutas
}
