% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOFishing.R
\name{readFAOFishing}
\alias{readFAOFishing}
\title{Read FAOSTAT FAOFishing}
\usage{
readFAOFishing()
}
\value{
A list containing the data as a magpie object and its metadata.
}
\description{
Reads the FAO projections CSV file and converts it into a magpie object.
}
\examples{
\dontrun{
x <- readSource("FAOFishing")
}

}
\author{
Fotis Sioutas
}
