% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOForestry.R
\name{readFAOForestry}
\alias{readFAOForestry}
\title{Read FAOSTAT FAOForestry}
\usage{
readFAOForestry(subset = c("Area harvested", "Stocks"))
}
\value{
A list containing the data as a magpie object and its metadata.
}
\description{
Reads the normalized FAOSTAT FAOForestry Sheets CSV file and converts
it into a magpie object.
}
\examples{
\dontrun{
x <- readSource("FAOForestry")
}

}
\author{
Fotis Sioutas
}
