% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOLandUse.R
\name{readFAOLandUse}
\alias{readFAOLandUse}
\title{Read FAOSTAT FAOLandUse}
\usage{
readFAOLandUse()
}
\value{
A list containing the data as a magpie object and its metadata.
}
\description{
Reads the FAO projections CSV file and converts it into a magpie object.
}
\examples{
\dontrun{
x <- readSource("FAOLandUse")
}

}
\author{
Fotis Sioutas
}
