% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOProductionCrops.R
\name{readFAOProductionCrops}
\alias{readFAOProductionCrops}
\title{Read FAOSTAT ProductionCrops}
\usage{
readFAOProductionCrops(subset = c("Area harvested", "Stocks"))
}
\value{
A list containing the data as a magpie object and its metadata.
}
\description{
Reads the normalized FAOSTAT ProductionCrops Sheets CSV file and converts
it into a magpie object.
}
\examples{
\dontrun{
x <- readSource("FAOProductionCrops")
}

}
\author{
Fotis Sioutas
}
