% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAOprojections.R
\name{readFAOprojections}
\alias{readFAOprojections}
\title{Read FAOSTAT projections}
\usage{
readFAOprojections()
}
\value{
A list containing the data as a magpie object and its metadata.
}
\description{
Reads the FAO projections CSV file and converts it into a magpie object.
}
\examples{
\dontrun{
x <- readSource("FAOprojections")
}

}
\author{
Fotis Sioutas
}
