% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readff55_Energy_Transport_GHG.R
\name{readff55_Energy_Transport_GHG}
\alias{readff55_Energy_Transport_GHG}
\title{readff55_Energy_Transport_GHG}
\usage{
readff55_Energy_Transport_GHG()
}
\value{
The read-in data into a magpie object.
}
\description{
Read energy-transport-ghg of EU
}
\examples{
\dontrun{
a <- readSource("ff55_Energy_Transport_GHG", convert = FALSE)
}

}
\author{
Fotis Sioutas
}
