% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readff55_Non_Co2_GHG_Emissions.R
\name{readff55_Non_Co2_GHG_Emissions}
\alias{readff55_Non_Co2_GHG_Emissions}
\title{readff55_Non_Co2_GHG_Emissions}
\usage{
readff55_Non_Co2_GHG_Emissions()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Non_Co2_GHG_Emissions of EU
}
\examples{
\dontrun{
a <- readSource("ff55_Non_Co2_GHG_Emissions", convert = FALSE)
}

}
\author{
Fotis Sioutas
}
