% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readff55EnergyTransportGHG.R
\name{readff55EnergyTransportGHG}
\alias{readff55EnergyTransportGHG}
\title{readff55EnergyTransportGHG}
\usage{
readff55EnergyTransportGHG()
}
\value{
The read-in data into a magpie object.
}
\description{
Read energy-transport-ghg of EU
}
\examples{
\dontrun{
a <- readSource("ff55EnergyTransportGHG", convert = FALSE)
}

}
\author{
Fotis Sioutas
}
