% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readff55NonCo2GHGEmissions.R
\name{readff55NonCo2GHGEmissions}
\alias{readff55NonCo2GHGEmissions}
\title{readff55NonCo2GHGEmissions}
\usage{
readff55NonCo2GHGEmissions()
}
\value{
The read-in data into a magpie object.
}
\description{
Read ff55NonCo2GHGEmissions of EU
}
\examples{
\dontrun{
a <- readSource("ff55NonCo2GHGEmissions", convert = FALSE)
}

}
\author{
Fotis Sioutas
}
