% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFullData.R
\name{readFullData}
\alias{readFullData}
\title{readFullData}
\usage{
readFullData(subtype)
}
\arguments{
\item{subtype}{Variable of MENA_EDS}
}
\value{
The read-in data into a magpie object.
}
\description{
Read MENA_EDS FullData gdx file
}
\examples{
\dontrun{
a <- readSource("FullData", subtype =  "VEH")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
