% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEM.R
\name{readGEM}
\alias{readGEM}
\title{readGEM}
\usage{
readGEM()
}
\value{
quitte object with the requested output data about
capacities of power plants from the Global Energy Monitor.
}
\description{
Read in a XLSX file and convert it to a quitte object
The data has information about capacities of power plants from the
Global Energy Monitor.
}
\examples{
\dontrun{
a <- readSource("GEM")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
