% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGidden.R
\name{readGidden}
\alias{readGidden}
\title{readGidden}
\usage{
readGidden()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Gidden Gt CO2 emissions.
}
\examples{
\dontrun{
a <- readSource("Gidden", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
