% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGlobalSolarAtlas.R
\name{readGlobalSolarAtlas}
\alias{readGlobalSolarAtlas}
\title{Read GlobalSolarAtlas data about solar potential per country.}
\usage{
readGlobalSolarAtlas()
}
\value{
The read-in data into a magpie object.
}
\description{
Read GlobalSolarAtlas data about solar potential per country.
}
\examples{
\dontrun{
a <- readSource("GlobalSolarAtlas", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
