% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGlobalWindAtlas.R
\name{readGlobalWindAtlas}
\alias{readGlobalWindAtlas}
\title{readGlobalWindAtlas}
\usage{
readGlobalWindAtlas()
}
\value{
magpie object with the requested output data about
wind potentials.
}
\description{
Read in wind powerDensity csv files from GlobalWindAtlas and convert it to a
magpie object. The data has information about wind potentials.
}
\examples{
\dontrun{
a <- readSource("GlobalWindAtlas")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
