% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGLOBIOMEU.R
\name{readGLOBIOMEU}
\alias{readGLOBIOMEU}
\title{readGLOBIOMEU}
\usage{
readGLOBIOMEU()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Emissions|CO2|AFOLU of all world countries from GLOBIOMEU
}
\examples{
\dontrun{
a <- readSource("GLOBIOMEU", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
