% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGLOBIUMEU_LAND.R
\name{readGLOBIOMEU_LAND}
\alias{readGLOBIOMEU_LAND}
\title{readGLOBIOMEU_LAND}
\usage{
readGLOBIOMEU_LAND()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Emissions|CO2|AFOLU|LAND of all EU countries from GLOBIOMEU
}
\examples{
\dontrun{
a <- readSource("GLOBIOMEU_LAND")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
