% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHotMaps.R
\name{readHotMaps}
\alias{readHotMaps}
\title{readHotMaps}
\usage{
readHotMaps()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Residential Heating Loads from Hotmaps
}
\examples{
\dontrun{
a <- readSource("HotMaps")
}

}
\author{
Margarita Efthymiadou
}
