% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readICTPrisma.R
\name{readICTPrisma}
\alias{readICTPrisma}
\title{readICTPrisma}
\usage{
readICTPrisma(subtype)
}
\value{
The read-in data into a magpie object.

 @param subtype By choosing a subtype you can select the corresponding csv
file and convert it to a magpie object.
Available types are:
\itemize{
\item `Number of data centers`:
\item `Consumption of data centers`:
}
}
\description{
readICTPrisma
}
\examples{
\dontrun{
a <- readSource("ICTPrisma")
}

}
\author{
Fotis Sioutas
}
