% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA.R
\name{readIEA}
\alias{readIEA}
\title{readIEA}
\usage{
readIEA(subtype = "MAINELEC")
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in energy balances from International Energy Agency.
}
\examples{
\dontrun{
a <- readSource("IEA", subtype = "MAINELEC")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
