% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA2025.R
\name{readIEA2025}
\alias{readIEA2025}
\title{readIEA2025}
\usage{
readIEA2025(subset = "INDPROD")
}
\arguments{
\item{subset}{flow : Type of data that should be read, e.g.
INDPROD :  production of primary energy or TOTENGY : Energy industry own use}
}
\value{
The read-in data into a magpie object.
}
\description{
The World Energy Balances data contains energy balances for 156 countries and
expressed in kilo tonnes of oil equivalent (ktoe). Conversion factors used
to calculate energy balances and indicators (including GDP, population,
industrial production index and ratios calculated with the energy data) are
also provided. The database also includes transparent notes on methodologies
and sources for country-level data. In general, the data are available from
1971 (1960 for OECD countries) to 2023. Preliminary 2023 data are available
for select countries, products, and flows.
}
\examples{
\dontrun{
a <- readSource("IEA2025", subtype = "INDPROD", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
