% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_CO2.R
\name{readIEA_CO2}
\alias{readIEA_CO2}
\title{readIEA_CO2}
\usage{
readIEA_CO2()
}
\value{
The read-in data into a magpie object.
}
\description{
Read IEA_CO2 2021 emissions from IEA.
}
\examples{
\dontrun{
a <- readSource("IEA_CO2", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
