% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_ENERGY_AI_Regional.R
\name{readIEA_ENERGY_AI_Regional}
\alias{readIEA_ENERGY_AI_Regional}
\title{readIEA_ENERGY_AI_Regional}
\usage{
readIEA_ENERGY_AI_Regional()
}
\value{
The read-in data into a magpie object.
}
\description{
ReadENERGY and AI data from IEA Regional.
}
\examples{
\dontrun{
a <- readSource("IEA_ENERGY_AI_Regional", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
