% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_ENERGY_AI_World.R
\name{readIEA_ENERGY_AI_World}
\alias{readIEA_ENERGY_AI_World}
\title{readIEA_ENERGY_AI_World}
\usage{
readIEA_ENERGY_AI_World()
}
\value{
The read-in data into a magpie object.
}
\description{
ReadENERGY and AI data from IEA for World.
}
\examples{
\dontrun{
a <- readSource("IEA_ENERGY_AI_World")
}

}
\author{
Fotis Sioutas
}
