% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/readIEA_Energy_End_uses_and_Efficiency_Indicators.R
\name{readIEA_Energy_End_uses_and_Efficiency_Indicators}
\alias{readIEA_Energy_End_uses_and_Efficiency_Indicators}
\title{readIEA_Energy_End_uses_and_Efficiency_Indicators}
\usage{
readIEA_Energy_End_uses_and_Efficiency_Indicators(subtype = "GENERIC")
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in IEA_Energy_End_uses_and_Efficiency_Indicators from International Energy Agency.
}
\examples{
\dontrun{
a <- readSource("IEA_Energy_End_uses_and_Efficiency_Indicators", subtype = "GENERIC")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
