% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_Energy_Projections_Balances.R
\name{readIEA_Energy_Projections_Balances}
\alias{readIEA_Energy_Projections_Balances}
\title{readIEA_Energy_Projections_Balances}
\usage{
readIEA_Energy_Projections_Balances(subtype = "INDPROD")
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in Energy_Projections_Balances from International Energy Agency.
}
\examples{
\dontrun{
a <- readSource("IEA_Energy_Projections_Balances", subtype = "INDPROD")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
