% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_Energy_Projections_Balances2025.R
\name{readIEA_Energy_Projections_Balances2025}
\alias{readIEA_Energy_Projections_Balances2025}
\title{readIEA_Energy_Projections_Balances2025}
\usage{
readIEA_Energy_Projections_Balances2025(subtype = "INDPROD")
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in Energy_Projections_Balances from International Energy Agency.
The period is: 2030, 2040 and 2050 based on Business as usual.
}
\examples{
\dontrun{
a <- readSource("IEA_Energy_Projections_Balances2025", subtype = "INDPROD")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
