% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_Energy_Projections_Indicators.R
\name{readIEA_Energy_Projections_Indicators}
\alias{readIEA_Energy_Projections_Indicators}
\title{readIEA_Energy_Projections_Indicators}
\usage{
readIEA_Energy_Projections_Indicators()
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in IEA_Energy_Projections_Indicators from International Energy Agency.
The period is: 2030, 2040 and 2050 based on Business as usual.
}
\examples{
\dontrun{
a <- readSource("IEA_Energy_Projections_Indicators")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
