% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_EV.R
\name{readIEA_EV}
\alias{readIEA_EV}
\title{readIEA_EV}
\usage{
readIEA_EV()
}
\value{
The read-in data into a magpie object.
}
\description{
Read IEA_EV data from IEA.
}
\examples{
\dontrun{
a <- readSource("IEA_EV", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
