% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_PRICES.R
\name{readIEA_PRICES}
\alias{readIEA_PRICES}
\title{readIEA_PRICES}
\usage{
readIEA_PRICES(subtype = "MAINELEC")
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in energy prices from International Energy Agency.
}
\examples{
\dontrun{
a <- readSource("IEA_PRICES", subtype = "ELECTR")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
