% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_WEO.R
\name{readIEA_WEO}
\alias{readIEA_WEO}
\title{readIEA_WEO}
\usage{
readIEA_WEO(subtype = "FuelPrices")
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read fossil fuel price projections from the IEA World Energy Outlook 2023
global source of energy analysis and projections.
The projections are based on the Stated Policies Scenario (STEPS).
It identifies and explores the biggest trends in energy demand and supply,
as well as what they mean for energy security, emissions and
economic development
}
\examples{
\dontrun{
a <- readSource("IEA_WEO", subtype = "FuelPrices")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
