% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_WEO_2023_ExtendedData.r
\name{readIEA_WEO_2023_ExtendedData}
\alias{readIEA_WEO_2023_ExtendedData}
\title{readIEA_WEO_2023_ExtendedData}
\usage{
readIEA_WEO_2023_ExtendedData(subtype = "IEA_WEO_2023_ExtendedData")
}
\arguments{
\item{subtype}{The folder name containing the data file.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in the WEO 2023 Extended Data Regions file for extracting industry production. More variables are included
}
\examples{
\dontrun{
a <- readSource("IEA_WEO_2023_ExtendedData", subtype = "IEA_WEO_2023_ExtendedData")
}

}
\author{
Sonja Sechi
}
