% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_WEO_TechCosts.R
\name{readIEA_WEO_TechCosts}
\alias{readIEA_WEO_TechCosts}
\title{readIEA_WEO_TechCosts}
\usage{
readIEA_WEO_TechCosts()
}
\value{
The read-in data into a magpie object.
}
\description{
Read PG_Assumptions TechCosts from the IEA World Energy Outlook 2023 global
source of energy analysis and projections.
The  Scenarios are Stated Policies and Net Zero Emissions by 2050.
Available types are: Renewables, Nuclear, Fossil fuels equipped with CCUS,
Coal, Gas and convert it to a magpie object.
}
\examples{
\dontrun{
a <- readSource("IEA_WEO_TechCosts")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
