% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEACrudeImportCountry.R
\name{readIEACrudeImportCountry}
\alias{readIEACrudeImportCountry}
\title{readIEACrudeImportCountry}
\usage{
readIEACrudeImportCountry()
}
\value{
The read-in data into a magpie object.
}
\description{
Read in Crude import per country from International Energy Agency.
}
\examples{
\dontrun{
a <- readSource("IEACrudeImportCountry")
}

}
\author{
Fotis Sioutas
}
