% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEACrudeOilPrice.R
\name{readIEACrudeOilPrice}
\alias{readIEACrudeOilPrice}
\title{readIEACrudeOilPrice}
\usage{
readIEACrudeOilPrice(subtype = "WORLD")
}
\value{
The read-in data into a magpie object.
}
\description{
Read in Crude Oil energy prices from International Energy Agency per sector, period, country.
}
\examples{
\dontrun{
a <- readSource("IEACrudeOilPrice")
}

}
\author{
Fotis Sioutas
}
