% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEACrudeSpotPrices.R
\name{readIEACrudeSpotPrices}
\alias{readIEACrudeSpotPrices}
\title{readIEACrudeSpotPrices}
\usage{
readIEACrudeSpotPrices()
}
\value{
The read-in data into a magpie object.
}
\description{
Read in Crude Spot Prices from International Energy Agency.
}
\examples{
\dontrun{
a <- readSource("IEACrudeSpotPrices")
}

}
\author{
Fotis Sioutas
}
