% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEAEnergyPrices.R
\name{readIEAEnergyPrices}
\alias{readIEAEnergyPrices}
\title{readIEAEnergyPrices}
\usage{
readIEAEnergyPrices(subtype = "RESID")
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in energy prices from International Energy Agency per sector, period, country and fuel.
}
\examples{
\dontrun{
a <- readSource("IEAEnergyPrices", subtype = "RESID")
}

}
\author{
Fotis Sioutas
}
