% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEATOTPRICES.R
\name{readIEATOTPRICES}
\alias{readIEATOTPRICES}
\title{readIEATOTPRICES}
\usage{
readIEATOTPRICES()
}
\value{
The read-in data into a magpie object.
}
\description{
Read in energy prices from International Energy Agency per sector, period, country and fuel.
}
\examples{
\dontrun{
a <- readSource("IEATOTPRICES")
}

}
\author{
Fotis Sioutas
}
