% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIMF.R
\name{readIMF}
\alias{readIMF}
\title{readIMF}
\usage{
readIMF()
}
\value{
The read-in data into a magpie object.
}
\description{
Read IMF GDP
}
\examples{
\dontrun{
a <- readSource("IMF", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
