% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIndustrialProcessesEmissionsIEAPrimes.R
\name{readIndustrialProcessesEmissionsIEAPrimes}
\alias{readIndustrialProcessesEmissionsIEAPrimes}
\title{readIndustrialProcessesEmissionsIEAPrimes}
\usage{
readIndustrialProcessesEmissionsIEAPrimes()
}
\value{
The read-in data into a magpie object.
}
\description{
Read IndustrialProcessesEmissions of all world countries from IEA Primes
}
\examples{
\dontrun{
a <- readSource("IndustrialProcessesEmissionsIEAPrimes")
}

}
\author{
Fotis Sioutas
}
