% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIRENA.R
\name{readIRENA}
\alias{readIRENA}
\title{readIRENA}
\usage{
readIRENA()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read IRENA
The dataset contains Capacity and generation data.
}
\examples{
\dontrun{
a <- readSource("IRENA")
}

}
\author{
Fotis Sioutas
}
