% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readiTEM.R
\name{readiTEM}
\alias{readiTEM}
\title{readiTEM}
\usage{
readiTEM()
}
\value{
The read-in data into a magpie object
}
\description{
Read in a csv file and convert it to a magpie object
Reads the International Transport Energy Modeling (iTEM) Open Data
https://zenodo.org/records/4287423#.ZCr0YOxBzX0
It contains various variables about transport, population and emissions
}
\examples{
\dontrun{
a <- readSource("iTEM")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
