% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLandAreaCountries.R
\name{readLandAreaCountries}
\alias{readLandAreaCountries}
\title{readLandAreaCountries}
\usage{
readLandAreaCountries()
}
\value{
magpie object with the requested output data about
LandAreaCountries.
}
\description{
Read in a csv file and convert it to a magpie object
The data has information about LandAreaCountries in sq per km2.
}
\examples{
\dontrun{
a <- readSource("LandAreaCountries")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
