% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLandlocked_Countries.R
\name{readLandlocked_Countries}
\alias{readLandlocked_Countries}
\title{readLandlocked_Countries}
\usage{
readLandlocked_Countries()
}
\value{
magpie object with the requested output data about
Landlocked Countries.
}
\description{
Read in a csv file and convert it to a magpie object
The data has information about Landlocked Countries.
}
\examples{
\dontrun{
a <- readSource("Landlocked_Countries")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
