% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLifetimesTranstech.R
\name{readLifetimesTranstech}
\alias{readLifetimesTranstech}
\title{readLifetimesTranstech}
\usage{
readLifetimesTranstech()
}
\value{
The read-in data into a magpie object.
}
\description{
Read in lifetimes for Transport sector
}
\examples{
\dontrun{
a <- readSource("LifetimesTranstech")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
