% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMACC.R
\name{readMACC}
\alias{readMACC}
\title{readMACC}
\usage{
readMACC()
}
\value{
A named list containing SSP2 CH4/N2O and SSP2 F-gas MAC curve
        data frames.
}
\description{
Read the CH4/N2O and F-gas MAC curve XLSX files and convert them into
structured data frames. The function extracts only sheets that contain
SSP2-compatible information (i.e., sheets whose names start with "SSP2"),
and returns them as a list of data frames.
}
\details{
The CH4/N2O dataset provides SSP2 baseline emissions (in Mt C-equivalent)
and marginal abatement cost (MAC) curves expressed as relative emission
reduction potentials. The F-gas dataset provides SSP2 baseline emissions
in kilotonnes of gas mass and MAC curves expressed as reductions relative
to baseline, along with costs in 2005 USD per tonne of C-equivalent.

The resulting data structure can be used directly in IAMs that apply
emission-based MAC curves rather than emission-factor-based approaches
}
\examples{
\dontrun{
a <- readSource("MACC")
}

}
\author{
Anastasis Giannousakis, Alexandros Tsimpoukis
}
