% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMAgPIE_BMSWAS_Price.R
\name{readMAgPIE_BMSWAS_Price}
\alias{readMAgPIE_BMSWAS_Price}
\title{readMAgPIE_BMSWAS_Price}
\usage{
readMAgPIE_BMSWAS_Price()
}
\value{
The MAgPIE BMSWAS price data as a magpie object.
}
\description{
Read annual regional bioenergy prices prepared from a MAgPIE report for use
as OPEN-PROM's historical BMSWAS fuel price. The source file is already in
OPEN-PROM region and subsector space and uses kUSD2015/toe.
}
\examples{
\dontrun{
x <- readSource("MAgPIE_BMSWAS_Price")
}

}
\author{
Songmin Yu
}
