% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMAGPIE_runs.R
\name{readMAGPIE_runs}
\alias{readMAGPIE_runs}
\title{readMAGPIE_runs}
\usage{
readMAGPIE_runs()
}
\value{
The read-in data into a magpie object.
}
\description{
Read the output of MAGPIE.
}
\examples{
\dontrun{
a <- readSource("MAGPIE_runs")
}

}
\author{
Fotis Sioutas
}
