% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMENA_EDS.R
\name{readMENA_EDS}
\alias{readMENA_EDS}
\title{readMENA_EDS}
\usage{
readMENA_EDS()
}
\value{
The read-in data into a mif object.
}
\description{
Read gdx files, convert it to a mif file so to compare MENA_EDS output mif file
with OPEN-PROM output.
}
\examples{
\dontrun{
a <- readSource("MENA_EDS")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
