% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNavigateConsumption.R
\name{readNavigateConsumption}
\alias{readNavigateConsumption}
\title{readNavigateConsumption}
\usage{
readNavigateConsumption()
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in NAVIGATE project model scenarios from IIASA.
The models are : 
COFFEE 1.3, REMIND-MAgPIE 3.0.0-4.5.0, IMAGE 3.2,PROMETHEUS V1,
GEM-E3_V2023,MESSAGEix-Materials,POLES NAVIGATE.
}
\details{
The scenarios are : 
NAV_Ind_NPi,NAV_Dem-NPi-ref.
}
\examples{
\dontrun{
a <- readSource("NavigateConsumption")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
