% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNDC_LTT_NECP.R
\name{readNDC_LTT_NECP}
\alias{readNDC_LTT_NECP}
\title{readNDC_LTT_NECP}
\usage{
readNDC_LTT_NECP()
}
\value{
A named list containing targets dataframes
}
\description{
Read the Global targets from NDCs and LTTs from XLSX file and convert them into
structured data frames.
}
\examples{
\dontrun{
a <- readSource("NDC_LTT_NECP")
}

}
\author{
Anastasis Giannousakis, Alexandros Tsimpoukis
}
