% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOPEN_GEM.R
\name{readOPEN_GEM}
\alias{readOPEN_GEM}
\title{readOPEN_GEM}
\usage{
readOPEN_GEM()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Production Level and Unit Cost data as delivered in GDX files from GEME3.
}
\examples{
\dontrun{
a <- readSource("OPEN_GEM")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
