% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPIK.R
\name{readPIK}
\alias{readPIK}
\title{readPIK}
\usage{
readPIK()
}
\value{
The read-in data into a magpie object.
}
\description{
Read PIK_CO2 emissions from PIK.
}
\examples{
\dontrun{
a <- readSource("PIK", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
