% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimes.R
\name{readPrimes}
\alias{readPrimes}
\title{readPrimes}
\usage{
readPrimes()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Primes data :
}
\examples{
\dontrun{
a <- readSource("Primes")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
