% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesBalances.R
\name{readPrimesBalances}
\alias{readPrimesBalances}
\title{readPrimesBalances}
\usage{
readPrimesBalances()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Primes data Final Energy Demand in DOMSE, INDSE, NENSE, TRANSE sector:
}
\examples{
\dontrun{
a <- readSource("PrimesBalances")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
