% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesBalancesGR.R
\name{readPrimesBalancesGR}
\alias{readPrimesBalancesGR}
\title{readPrimesBalancesGR}
\usage{
readPrimesBalancesGR(subtype)
}
\value{
The read-in data into a magpie object.
}
\description{
readPrimesBalancesGR
}
\examples{
\dontrun{
a <- readSource("PrimesBalancesGR")
}

}
\author{
Fotis Sioutas
}
