% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesFuelPrices.R
\name{readPrimesFuelPrices}
\alias{readPrimesFuelPrices}
\title{readPrimesFuelPrices}
\usage{
readPrimesFuelPrices()
}
\value{
The read-in data into a magpie object.
}
\description{
readPrimesFuelPrices
}
\examples{
\dontrun{
a <- readSource("PrimesFuelPrices")
}

}
\author{
Fotis Sioutas
}
