% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesNewTransport.R
\name{readPrimesNewTransport}
\alias{readPrimesNewTransport}
\title{readPrimesNewTransport}
\usage{
readPrimesNewTransport(subtype = "FuelOutlook")
}
\arguments{
\item{subtype}{excel sheet (FuelOutlook, Stock or Indicators)}
}
\value{
The read-in data into a magpie object.
}
\description{
Read Primes data:
}
\examples{
\dontrun{
a <- readSource("PrimesNewTransport", subtype = "FuelOutlook")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
