% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesPGData.R
\name{readPrimesPGData}
\alias{readPrimesPGData}
\title{readPrimesPGData}
\usage{
readPrimesPGData(esxel_sheet)
}
\arguments{
\item{esxel_sheet}{excel sheet "Summary" for SEElec or "power generation" for
Gross Electricity generation by fuel type}
}
\value{
The read-in data into a magpie object.
}
\description{
Read Primes Secondary Energy Electricity Gross and Gross Electricity
generation by fuel type for EU countries:
}
\examples{
\dontrun{
a <- readSource("PrimesPGData", esxel_sheet = "Summary")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
