% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesPrices.R
\name{readPrimesPrices}
\alias{readPrimesPrices}
\title{readPrimesPrices}
\usage{
readPrimesPrices()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Primes Prices per fuel and sector for EU countries:
}
\examples{
\dontrun{
a <- readSource("PrimesPrices")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
