% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPRIMESPROCTECH.R
\name{readPRIMESPROCTECH}
\alias{readPRIMESPROCTECH}
\title{readPRIMESPROCTECH}
\usage{
readPRIMESPROCTECH()
}
\value{
The read-in data into magpie object.
}
\description{
Read PRIMES tech data and convert it to a magpie object
}
\examples{
\dontrun{
a <- readSource("PRIMESPROCTECH")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
