% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesSEElec.R
\name{readPrimesSEElec}
\alias{readPrimesSEElec}
\title{readPrimesSEElec}
\usage{
readPrimesSEElec()
}
\value{
The read-in data into a magpie object.
}
\description{
Read Primes Prices per fuel and sector for EU countries:
}
\examples{
\dontrun{
a <- readSource("PrimesSEElec")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
