% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesTransport.R
\name{readPrimesTransport}
\alias{readPrimesTransport}
\title{readPrimesTransport}
\usage{
readPrimesTransport()
}
\value{
The read-in data into a magpie object.
}
\description{
Read PrimesTransport data :
}
\examples{
\dontrun{
a <- readSource("PrimesTransport")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
