% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesTransportNew.R
\name{readPrimesTransportNew}
\alias{readPrimesTransportNew}
\title{readPrimesTransportNew}
\usage{
readPrimesTransportNew(subtype = "FuelOutlook")
}
\arguments{
\item{subtype}{excel sheet (FuelOutlook, Stock or Indicators)}
}
\value{
The read-in data into a magpie object.
}
\description{
Read PrimesTransportNew data :
}
\examples{
\dontrun{
a <- readSource("PrimesTransportNew", subtype = "FuelOutlook")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
