% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProjectCarbonPrices.R
\name{readProjectCarbonPrices}
\alias{readProjectCarbonPrices}
\title{readProjectCarbonPrices}
\usage{
readProjectCarbonPrices()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read Project Carbon Prices
The dataset contains carbon-price trajectories from several projects.
The csv file may be updated with new projects in the future.
Please have the values as US$2015/tCO2 otherwise convert them to [calcIEnvPolicies()].
}
\examples{
\dontrun{
a <- readSource("ProjectCarbonPrices")
}

}
\author{
Alexandros Tsimpoukis
}
