% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPROMParameters.R
\name{readPROMParameters}
\alias{readPROMParameters}
\title{readPROMParameters}
\usage{
readPROMParameters(subtype = "PriceTransElast")
}
\arguments{
\item{subtype}{One of:
"PriceTransElast" - fuel price pass-through elasticity, source fuel -> target fuel}
}
\value{
magpie object with the requested parameter table
}
\description{
Read hand-maintained OPEN-PROM / PROMETHEUS model parameters (behavioural
response coefficients: elasticities, response shares) from the PROMParameters
madrat source folder. One subtype per parameter table.
}
\examples{
\dontrun{
a <- readSource("PROMParameters", subtype = "PriceTransElast")
}

}
\author{
Songmin Yu
}
