% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMIND_MAgPIE_PRISMA.R
\name{readREMIND_MAgPIE_PRISMA}
\alias{readREMIND_MAgPIE_PRISMA}
\title{readREMIND_MAgPIE_PRISMA}
\usage{
readREMIND_MAgPIE_PRISMA()
}
\value{
The read-in data into a magpie object.
}
\description{
Read REMIND_MAgPIE_PRISMA emissions.
}
\examples{
\dontrun{
a <- readSource("REMIND_MAgPIE_PRISMA")
}

}
\author{
Christos Koumparakis
}
