% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readREMIND_MAgPIE_SoCDR.R
\name{readREMIND_MAgPIE_SoCDR}
\alias{readREMIND_MAgPIE_SoCDR}
\title{readREMIND_MAgPIE_SoCDR}
\usage{
readREMIND_MAgPIE_SoCDR()
}
\value{
The read-in data into a magpie object.
}
\description{
Read REMIND_MAgPIE_SoCDR emissions.
}
\examples{
\dontrun{
a <- readSource("REMIND_MAgPIE_SoCDR")
}

}
\author{
Fotis Sioutas
}
