% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSets.R
\name{readSets}
\alias{readSets}
\title{readSets}
\usage{
readSets(file, sector)
}
\arguments{
\item{file}{A gams file containing GAMS code.}

\item{sector}{A set from gams file.}
}
\value{
The values of set.
}
\description{
Reads all set values given in a GAMS code and returns them.
}
\examples{
\dontrun{
a <- readSets(file,"SBS")
}

}
\author{
Anastasis Giannousakis Fotis Sioutas
}
