% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSSPold.R
\name{readSSPold}
\alias{readSSPold}
\title{readSSPold}
\usage{
readSSPold()
}
\value{
magpie object object with the requested output GDP|PPP and Population
}
\description{
Read in a csv file and convert it to a magpie object
The data has information about GDP|PPP and Population.
}
\examples{
\dontrun{
a <- readSource("SSPold")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
