% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTDOMSEshareproj.R
\name{readTDOMSEshareproj}
\alias{readTDOMSEshareproj}
\title{readTDOMSEshareproj}
\usage{
readTDOMSEshareproj(subtype)
}
\arguments{
\item{subtype}{string.}
}
\value{
magpie object
}
\description{
readTDOMSEshareproj
}
\examples{
\dontrun{
TDOMSE <- readSource("TDOMSEshareproj", subtype = "Shares")
}

}
\author{
Fotis Sioutas
}
