% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTechCosts.R
\name{readTechCosts}
\alias{readTechCosts}
\title{readTechCosts}
\usage{
readTechCosts(subtype = "PowerAndHeat")
}
\arguments{
\item{subtype}{A character string referring to the excel sheet, e.g. "PowerAndHeat" which would
read the excel sheet power_and_heat_energyf from the file TechCosts and convert it to a magpie object}
}
\value{
The read-in data into a magpie object
}
\description{
Read in an excel file and convert it to a magpie object
The file containing the data is named TechCosts
}
\author{
Anastasis Giannousakis, Fotis Sioutas

#' @examples
\dontrun{
a <- readSource("TechCosts")
}
}
