% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTechCosts2024.R
\name{readTechCosts2024}
\alias{readTechCosts2024}
\title{readTechCosts2024}
\usage{
readTechCosts2024(subtype = "PowerAndHeat")
}
\arguments{
\item{subtype}{Type of data that should be read. The type is referring to the
excel sheet, from the excel file "REF2020_Technology Assumptions_Transport.xlsx"
and convert it to a magpie object.
Available types are:
\itemize{
\item `PowerAndHeat`:
\item `PowerAndHeatEfficiency`:
\item `DomesticEnergy`:
\item `IndustryEnergy`:
\item `infrastructure`:
\item `new_fuels_energy`:
\item `maritime`:
\item `Inland_navigation`:
\item `Rail`:
\item `Aviation`:
\item `2wheelers`:
\item `Bus_coach`:
\item `HGVs>16t`:
\item `HGVs<16t`:
\item `LCVs`:
\item `Large_cars`:
\item `Medium_cars`:
\item `Small_cars`:
\item `renovation_costs`:
\item `NewFuelsEnergyEFF`:
}}
}
\value{
The read-in data into a magpie object
}
\description{
Read in technology cost and other data from the "EU Reference Scenario":
https://energy.ec.europa.eu/data-and-analysis/energy-modelling/eu-reference-scenario-2020_en
and convert it to a magpie object
}
\examples{
\dontrun{
a <- readSource("TechCosts2024")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
