% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTSharesINDSE2.R
\name{readTINDSE2}
\alias{readTINDSE2}
\title{readTINDSE2}
\usage{
readTINDSE2()
}
\arguments{
\item{subtype}{string.}
}
\value{
magpie object
}
\description{
readTINDSE2
}
\examples{
\dontrun{
TTINDSE2 <- readSource("TINDSE2")
}

}
\author{
Fotis Sioutas
}
