% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTREMOVE.R
\name{readTREMOVE}
\alias{readTREMOVE}
\title{readTREMOVE}
\usage{
readTREMOVE(subtype = "Activity")
}
\arguments{
\item{subtype}{Type of data that should be read. The type is referring to the
excel sheet, and convert it to a magpie object.
Available types are:
\itemize{
\item `Activity`:
\item `Stock`:
\item `FuelOutlook`:
}}
}
\value{
The read-in data into a magpie object.
}
\description{
Read TREMOVE data :
Reference has national policies with accepted measures until end-2019. REF2020_v6
Mix is a decarb scenario contributing to the EU Green Deal goals towards net zero in 2050. MIX_2021_v1
}
\examples{
\dontrun{
a <- readSource("TREMOVE", subtype = "Activity")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
