% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTSharesDOMSE.R
\name{readTSharesDOMSE}
\alias{readTSharesDOMSE}
\title{readTSharesDOMSE}
\usage{
readTSharesDOMSE()
}
\value{
The read-in targets shares data into a magpie object
}
\description{
The dataset contains targets shares for DOMSE.
}
\examples{
\dontrun{
a <- readSource("TSharesDOMSE")
}

}
\author{
Fotis Sioutas
}
