% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTSharesINDSE.R
\name{readTSharesINDSE}
\alias{readTSharesINDSE}
\title{readTSharesINDSE}
\usage{
readTSharesINDSE(subtype)
}
\value{
The read-in targets shares data into a magpie object
}
\description{
The dataset contains targets shares for INDSE.
}
\examples{
\dontrun{
a <- readSource("TSharesINDSE")
}

}
\author{
Fotis Sioutas
}
