% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUN.R
\name{readUN}
\alias{readUN}
\title{readUN}
\usage{
readUN()
}
\value{
The read-in data into a magpie object
}
\description{
Read in a csv file and convert it to a magpie object
Reads the UN Open Data Energy Balances
https://data.un.org/SdmxBrowser/start
It contains various variables about balances, consumption, production, losses,
imports, exports and other.
}
\examples{
\dontrun{
a <- readSource("UN")
}

}
\author{
Fotis Sioutas
}
