% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUN_GBR_LULUCF.R
\name{readUN_GBR_LULUCF}
\alias{readUN_GBR_LULUCF}
\title{readUN_GBR_LULUCF}
\usage{
readUN_GBR_LULUCF()
}
\value{
The read-in data into a magpie object.
}
\description{
Read United Nations data for UK for LULUCF CO2 equivalent emissions.
}
\examples{
\dontrun{
a <- readSource("UN_GBR_LULUCF", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
