% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNFCCC.R
\name{readUNFCCC}
\alias{readUNFCCC}
\title{readUNFCCC}
\usage{
readUNFCCC(subtype = "2.  Industrial Processes and Product Use")
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in UNFCCC country-submitted greenhouse gas emissions data until 2024-07-05
}
\examples{
\dontrun{
a <- readSource("UNFCCC", subtype = "2.  Industrial Processes and Product Use")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
