% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUNPopDiv.r
\name{readUNPopDiv}
\alias{readUNPopDiv}
\title{readUNPopDiv}
\usage{
readUNPopDiv(subtype = "Bangladesh")
}
\value{
The read-in data into a magpie object
}
\description{
Read data (historical + estimates for medium variant) from the UN World Population Prospects 2024
}
\examples{
\dontrun{
a <- readSource("UNPopDiv", subtype =  "Bangladesh", convert = FALSE)
}
}
\author{
Anastasis Giannousakis, Fotis Sioutas, Alexandros Tsimpoukis
}
