% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUPTCarbonPrices.R
\name{readUPTCarbonPrices}
\alias{readUPTCarbonPrices}
\title{readUPTCarbonPrices}
\usage{
readUPTCarbonPrices()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read UPT Carbon Prices
The dataset contains UPT protocol carbon-price trajectories. Projection
values from 2023 onward are expressed in US$2010/tCO2; conversion to the
OPEN-PROM internal US$2015/tCO2 convention is handled by
[calcIEnvPolicies()].
}
\examples{
\dontrun{
a <- readSource("UPTCarbonPrices")
}

}
\author{
Fotis Sioutas
}
