% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUSGS.R
\name{readUSGS}
\alias{readUSGS}
\title{readUSGS}
\usage{
readUSGS()
}
\value{
The read-in data into a magpie object.
}
\description{
Read USGS RAW STEEL: WORLD PRODUCTION BY COUNTRY OR LOCALITY1, 2
}
\examples{
\dontrun{
a <- readSource("USGS", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
