% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWDI_PA.R
\name{readWDI_PA}
\alias{readWDI_PA}
\title{readWDI_PA}
\usage{
readWDI_PA()
}
\value{
magpie object with the requested output data about air transport passengers per country and per year
}
\description{
Read in a csv file and convert it to a magpie object
The data has information about air transport passengers per country and per year
}
\author{
Anastasis Giannousakis, Fotis Sioutas

#' @examples
\dontrun{
a <- readSource("WDI_PA")
}
}
