% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWEO2023CarbonPrices.R
\name{readWEO2023CarbonPrices}
\alias{readWEO2023CarbonPrices}
\title{readWEO2023CarbonPrices}
\usage{
readWEO2023CarbonPrices()
}
\value{
The read-in data into a magpie object.
}
\description{
Read WEOCarbonPrices from IEA.
}
\examples{
\dontrun{
a <- readSource("WEO2023CarbonPrices")
}

}
\author{
Fotis Sioutas
}
