% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWEO2023CarbonPrices.R
\name{readWEOCarbonPrices}
\alias{readWEOCarbonPrices}
\title{readWEOCarbonPrices}
\usage{
readWEOCarbonPrices()
}
\value{
The read-in data into a magpie object.
}
\description{
Read WEOCarbonPrices from IEA.
}
\examples{
\dontrun{
a <- readSource("WEOCarbonPrices", convert = TRUE)
}

}
\author{
Fotis Sioutas
}
