% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWEPP2022.R
\name{readWEPP2022}
\alias{readWEPP2022}
\title{readWEPP2022}
\usage{
readWEPP2022(subtype = "PV")
}
\arguments{
\item{subtype}{string. By choosing a subtype you filter the WEPP2022 dataset
(UTYPE column), to allow further processing of specific variables.
Available types are:
\itemize{
\item `PV`: Photovoltaic cells
\item `IC`: Internal combustion (reciprocating diesel) engine
\item `ST/S`: Steam turbine with steam sendout (cogen)
\item `IC/H`: Internal combustion engine with heat recovery (cogen - CHP)
\item `GT`: Gas/combustion turbine in simple (open) cycle
\item `HY`: Hydroelectric turbine generator
\item `WTG`: Wind turbine generator
\item `GT/H`: Microturbine with exhaust heat recovery (cogen - CHP)
\item `GT/S`: Gas turbine with steam sendout or other heat recovery (cogen - CHP)
\item `ST`: Steam turbine
\item `GT/C`: Gas turbine in combined-cycle
\item `ST/C`: Steam turbine in combined-cycle
\item `CC`: Combined-cycle, unspecified configuration
\item `ORC`: Organic Rankine-cycle (vapor) turbine or ORC energy converter, also includes some turbo-expanders
\item `PV/F`: Floating photovoltaic array
\item `WTG/O`: Wind turbine generator located offshore
\item `TTG`: Tidal or wave turbine generator or riverine kinetic energy system
\item `CC/S`: Combined-cycle in CHP (cogen) application with steam sendout or heat recovery
\item `CCSS`: Combined-cycle single shaft configuration
\item `GT/CP`: Gas turbine in combined-cycle CHP (cogen) power plant
\item `ST/CP`: Steam turbine in combined-cycle CHP (cogen) power plant
\item `FC`: Fuel cell
\item `TEX`: Turboexpander, gas expander or pressure recovery turbine
\item `IC/CD`: Internal combustion (reciprocating diesel) engine in combined-cycle
\item `ST/CD`: Steam turbine in combined-cycle with gas or diesel engines
\item `CCSS/P`: Combined-cycle single shaft configuration in CHP (cogen) application
\item `RSE`: Reciprocating steam engine
\item `FC/H`: Fuel cell with heat recovery
\item `GT/T`: Gas turbine in topping configuration with existing conventional boiler and T/G set
\item `TRT`: Top gas pressure recovery turbine, also known as a gas expansion turbine (used in steel works)
\item `IC/CP`: Internal combustion (reciprocating or diesel) engine in combined-cycle CHP (cogen) power plant
\item `WTG/OF`: Offshore wind turbine on floating platform
\item `ST/T`: Steam turbine with heat input from topping gas turbine
\item `PVH`: PV with integrated heat recovery, PV cogen
\item `IC/D`: Internal combustion engine with heat recovery for desalination
\item `ST/s`: Steam turbine with steam sendout (cogen)
\item `OTEC`: Ocean thermal energy conversion device
\item `ST/D`: Steam turbine with heat recovery for desalination
\item `ECE`: External combustion engine, Stirling or other designs, multifuel
\item `IFGT`: Indirectly-fired gas turbine, uses heat exchanger and operates as a hot air engine
\item `GT/R`: Gas turbine employed for steam-turbine repowering
\item `CC/CP`: Combined-cycle in CHP (cogen) application
\item `RSEX`: Rotary steam expander
\item `ORC/H`: ORC with heat recovery
\item `ACT`: Allam Cycle, oxy-combustion technology
\item `ST/CS`: Steam turbine in combined-cycle CHP (cogen) power plant for desalination
\item `GT/CS`: Gas turbine in combined-cycle CHP (cogen) power plant for desalination
\item `GT/D`: Gas turbine with heat recovery for desalination
\item `CC/D`: Combined-cycle with heat recovery for desalination
\item `SUN`: Solar power
}}
}
\value{
The read-in data into a magpie object
}
\description{
Read in an excel file and convert it to a magpie object
The data has information about electric power plant
}
\examples{
\dontrun{
a <- readSource("WEPP2022", subtype = "PV")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
