% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWorldBankCarPr.R
\name{readWorldBankCarPr}
\alias{readWorldBankCarPr}
\title{readWorldBankCarPr}
\usage{
readWorldBankCarPr()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read in data from the World Bank project.
The dataset contains carbon price data.
}
\examples{
\dontrun{
a <- readSource("WorldBankCarPr")
}

}
\author{
Anastasis Giannousakis
}
