% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWorldBankCarPr2025.R
\name{readWorldBankCarPr2025}
\alias{readWorldBankCarPr2025}
\title{readWorldBankCarPr2025}
\usage{
readWorldBankCarPr2025()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read in data from the World Bank project.
The dataset contains carbon price data.
}
\examples{
\dontrun{
a <- readSource("WorldBankCarPr2025")
}

}
\author{
Anastasis Giannousakis
}
