% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWorldBankWDI.R
\name{readWorldBankWDI}
\alias{readWorldBankWDI}
\title{readWorldBankWDI}
\usage{
readWorldBankWDI(subtype = "BGD")
}
\value{
magpie object with the requested output data about GDP per selected country and per year
}
\description{
Use library "WDI" to extract data from the World Bank and convert it to a magpie object
The data has information about GDP per country and per year. The indicator list is pre-defined and hardcoded. 
          indicators= "NY.GDP.MKTP.KD,NY.GDP.PCAP.KD,NY.GDP.DEFL.ZS,NY.GDP.MKTP.KD.ZG"  # GDP (constant 2015 US$), GDP per capita (constant 2015 US$), GDP deflator, annual growth
}
\examples{
\dontrun{
a <- readSource("WorldBankWDI",subtype ="BGD", convert = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Alexandros Tsimpoukis
}
