% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWSA.R
\name{readWSA}
\alias{readWSA}
\title{readWSA}
\usage{
readWSA()
}
\value{
The read-in data into a magpie object.
}
\description{
Read WSA.
}
\examples{
\dontrun{
a <- readSource("WSA")
}

}
\author{
Fotis Sioutas
}
