% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWSA_2025.R
\name{readWSA_2025}
\alias{readWSA_2025}
\title{readWSA_2025}
\usage{
readWSA_2025()
}
\value{
The read-in data into a quitte object.
}
\description{
Read World Steel Association Crude Steel production by process 2024
}
\examples{
\dontrun{
a <- readSource("WSA_2025")
}

}
\author{
Fotis Sioutas
}
