% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolComtradrToGDX.R
\name{toolComtradrToGDX}
\alias{toolComtradrToGDX}
\title{toolComtradrToGDX}
\usage{
toolComtradrToGDX(subtypes = "854143", start = "2011", end = "2022")
}
\arguments{
\item{subtypes}{Type of data.}

\item{start}{Starting date of data.}

\item{end}{Ending date of data.}
}
\value{
The read-in data into a gdx file
}
\description{
Read in data from the Comtradr and convert it to gdx file.
}
\examples{
\dontrun{
a <- toolComtradrToGDX(subtypes = c("854143","540110"), start = "2011", end = "2022")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
