% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFixWithRegional.R
\name{toolFixWithRegional}
\alias{toolFixWithRegional}
\title{toolFixWithRegional}
\usage{
toolFixWithRegional(x, h12)
}
\arguments{
\item{x}{MAgPIE object with ISO country codes.}

\item{h12}{type of region mapping.}
}
\value{
magpie object with the requested output data about regional 
mean value.
}
\description{
For the given MAgPIE object with ISO country codes, this function returns 
a MAgPIE object with the regional mean value in which each country is,
for the specific region, period and variable. 
If there is no data for the country the region value is taken.
If there is no data for the region the global value is taken.
}
\examples{
\dontrun{
a <- toolFixWithRegional(x, h12)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
