% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFunnelPlots.R
\name{toolFunnelPlots}
\alias{toolFunnelPlots}
\title{toolFunnelPlots}
\usage{
toolFunnelPlots(subtypes = "World")
}
\arguments{
\item{subtypes}{Region}
}
\value{
The read-in data to plots
}
\description{
Read in data XLSX file from IAMCOMPACT and make plots for "Final Energy" and
"Emissions|CO2" for all the studies. ENERDATA and NAVIGATE are used for validation.
}
\examples{
\dontrun{
a <- toolFunnelPlots(subtype = "World")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
