% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolIEAToGDX.R
\name{toolIEAToGDX}
\alias{toolIEAToGDX}
\title{toolIEAToGDX}
\usage{
toolIEAToGDX(subtype = "1960")
}
\arguments{
\item{subtype}{The year.}
}
\value{
The read-in data into a gdx file
}
\description{
toolIEAToGDX
}
\examples{
\dontrun{
a <- toolIEAToGDX(subtype = c("1960","1970"))
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
