% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolILOToGDX.R
\name{toolILOToGDX}
\alias{toolILOToGDX}
\title{toolILOToGDX}
\usage{
toolILOToGDX(subtypes = "POP_XWAP_SEX_AGE_NB_A")
}
\arguments{
\item{subtypes}{Type of data.}
}
\value{
The read-in data into a gdx file
}
\description{
Read in data from ILO and convert it to gdx file.
}
\examples{
\dontrun{
a <- toolILOToGDX(subtypes = c("POP_XWAP_SEX_AGE_NB_A", "EMP_TEMP_SEX_AGE_OCU_NB_A"))
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
