% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolOECDToGDX.R
\name{toolOECDToGDX}
\alias{toolOECDToGDX}
\title{toolOECDToGDX}
\usage{
toolOECDToGDX(subtypes = "all")
}
\arguments{
\item{subtypes}{Type of data.}
}
\value{
The read-in data into a gdx file
}
\description{
Read in data from the OECD and convert it to gdx file.
}
\examples{
\dontrun{
a <- toolOECDToGDX(subtypes = "IRS")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
