% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolRawDataTable.R
\name{toolRawDataTable}
\alias{toolRawDataTable}
\title{toolRawDataTable}
\usage{
toolRawDataTable(file)
}
\arguments{
\item{file}{The input.gms file path.}
}
\value{
Info about sources of mrprom.
}
\description{
Description of the sources of mrprom.
}
\examples{
\dontrun{
a <- toolRawDataTable(file)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
