% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReadEvalGlobal.R
\name{toolReadEvalGlobal}
\alias{toolReadEvalGlobal}
\title{toolReadEvalGlobal}
\usage{
toolReadEvalGlobal(file)
}
\arguments{
\item{file}{A gams file containing GAMS code.}
}
\value{
A vector of values the evalglobal variables are set to as names.
}
\description{
Reads all evalglobals given in a GAMS code and returns them.
}
\examples{
\dontrun{
toolReadEvalGlobal <- function(main.gms)
}


}
\author{
Anastasis Giannousakis Fotis Sioutas
}
