% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReadSetFromGDX.R
\name{toolReadSetFromGDX}
\alias{toolReadSetFromGDX}
\title{toolReadSetFromGDX}
\usage{
toolReadSetFromGDX()
}
\value{
Dataframe with the values_parameters of some variables
from fulldata.
}
\description{
For the given fulldata gdx file, this function returns a dataframe with the
values_parameters of some variables from fulldata.
}
\examples{
\dontrun{
a <- toolReadSetFromGDX()
}

}
\author{
Anastasis Giannousakis Fotis Sioutas
}
