% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolreadSets.R
\name{toolreadSets}
\alias{toolreadSets}
\title{toolreadSets}
\usage{
toolreadSets(file, sector)
}
\arguments{
\item{file}{The sets.gms file containing GAMS code.}

\item{sector}{A set from gams file.}
}
\value{
The values of set.
}
\description{
Reads all set values given in file sets.gms of OPEN-PROM and returns them.
}
\examples{
\dontrun{
a <- toolreadSets(file,"SBS")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
