% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSet.R
\name{toolSet}
\alias{toolSet}
\title{toolSet}
\usage{
toolSet(x, set)
}
\arguments{
\item{x}{dataframe}

\item{set}{column of (x)}
}
\value{
A list in order to create a gdx file.
}
\description{
For a given dataframe (x), this function returns
a list with the unique values of the column set
and with the standard format in order to create a gdx file.
}
\examples{
\dontrun{
a <- toolSet(x, "region")
}


}
\author{
Anastasis Giannousakis Fotis Sioutas
}
