% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolUNtoGDX.R
\name{toolUNtoGDX}
\alias{toolUNtoGDX}
\title{toolUNtoGDX}
\usage{
toolUNtoGDX(subtypes = "total")
}
\arguments{
\item{subtypes}{Type of aggregation.}
}
\value{
The read-in data into a gdx file
}
\description{
Read in data from the UN and convert it to gdx file.
}
\examples{
\dontrun{
a <- toolUNtoGDX()
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
