% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcACTV.R
\name{calcACTV}
\alias{calcACTV}
\title{calcACTV}
\usage{
calcACTV()
}
\value{
The read-in data into a magpie object.
}
\description{
Derive economic activity data for OPENPROM sectors based on two data sources:
transport, traffic, air transport passengers per country and per year (IRF)
and Production Level and Unit Cost (GEME3).
}
\examples{
\dontrun{
a <- calcOutput(type = "ACTV", file = "iACTV.csv", aggregate = TRUE)
}
}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
