% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcACTV.R
\name{calcACTV}
\alias{calcACTV}
\title{calcACTV}
\usage{
calcACTV()
}
\value{
The read-in data into a magpie object.
}
\description{
Read in several files with data from GEME3, IRF and WDI_PA.
The dataset contains several data types about transport, traffic, air
transport passengers per country and per year and Production Level and
Unit Cost data.
}
\examples{
\dontrun{
a <- calcOutput(type = "ACTV", file = "iACTV.csv", aggregate = TRUE)
}
}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
