% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcElectrProd.R
\name{calcElectrProd}
\alias{calcElectrProd}
\title{calcElectrProd}
\usage{
calcElectrProd()
}
\value{
Derive OPENPROM input data on electricity production from ENERDATA.
}
\description{
calcElectrProd
}
\examples{
\dontrun{
a <- calcOutput(type = "ElectrProd", file = "iElectrProd.csvr")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
