% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFIT.R
\name{calcFIT}
\alias{calcFIT}
\title{calcFIT}
\usage{
calcFIT()
}
\value{
Magpie object with the Feed-In Tariff (FIT) for electricity per year
}
\description{
Calculate the Feed-In Tariff (FIT) for electricity per year
}
\examples{
\dontrun{
a <- calcOutput(type = "FIT", aggregate = FALSE)
}

}
\author{
Michalis Madianos, Dionysis Pramangioulis
}
