% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFuelConsumption.R
\name{calcFuelConsumption}
\alias{calcFuelConsumption}
\title{calcFuelConsumption}
\usage{
calcFuelConsumption(subtype = "transport")
}
\arguments{
\item{subtype}{string. Specific type of consumption.}
}
\value{
The "ENERDATA" data (filtered by consumption and type of consumption).
}
\description{
The two available ENERDATA datasets are combined into one filtered by
consumption, and allow further processing of a specific type of consumption.
}
\examples{
\dontrun{
a <- calcOutput(type = "FuelConsumption", file = "transport.csv" , aggregate = FALSE)
}


}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
