% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIAvailRate.R
\name{calcIAvailRate}
\alias{calcIAvailRate}
\title{calcIAvailRate}
\usage{
calcIAvailRate()
}
\value{
magpie object with OPENPROM input data iAvailRate
}
\description{
Use data from EU Reference Scenario to derive OPENPROM input parameter iAvailRate
This dataset includes plant availability rate, as a percentage.
}
\examples{
\dontrun{
a <- calcOutput(type = "IAvailRate", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
