% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIChpPowGen.R
\name{calcIChpPowGen}
\alias{calcIChpPowGen}
\title{calcIChpPowGen}
\usage{
calcIChpPowGen()
}
\value{
magpie object with OPENPROM input data iChpPowGen
}
\description{
Use data from EU Reference Scenario to derive OPENPROM input table imDataChpPowGen
This dataset includes CHP economic and technical data initialisation for electricity production.
The availability factor is hardcoded and it is based on previous data.
}
\examples{
\dontrun{
a <- calcOutput(type = "IChpPowGen", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Alexandros Tsimpoukis
}
