% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataAgricultureService.R
\name{calcIDataAgricultureService}
\alias{calcIDataAgricultureService}
\title{calcIDataAgricultureService
Units : Crops - 1e9 ha, Livestock - 1e9 An, Forestry - 1e6 m3,
        Fishing - 1e6 tonnes, IRRIGATION - 1e9 ha}
\usage{
calcIDataAgricultureService()
}
\value{
Magpie object with the FAOProductionCrops
}
\description{
calcIDataAgricultureService
Units : Crops - 1e9 ha, Livestock - 1e9 An, Forestry - 1e6 m3,
        Fishing - 1e6 tonnes, IRRIGATION - 1e9 ha
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataAgricultureService", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
