% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataConsEneBranch.R
\name{calcIDataConsEneBranch}
\alias{calcIDataConsEneBranch}
\title{calcIDataConsEneBranch}
\usage{
calcIDataConsEneBranch()
}
\value{
magpie object with OPENPROM input data iDataConsEneBranch.
}
\description{
Use data from ENERDATA to derive OPENPROM input parameter iDataConsEneBranch.
This dataset includes consumption values for each region and energy branch in Mtoe.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataConsEneBranch", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
