% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataDistrLosses.R
\name{calcIDataDistrLosses}
\alias{calcIDataDistrLosses}
\title{calcIDataDistrLosses}
\usage{
calcIDataDistrLosses()
}
\value{
OPENPROM input data iDataDistrLosses
}
\description{
Transport distribution losses for Natural Gas, Other Gases and Electricity
	from ENERDATA.
	Use data to derive OPENPROM input parameter iDataDistrLosses
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataDistrLosses", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
