% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataElecInd.R
\name{calcIDataElecInd}
\alias{calcIDataElecInd}
\title{calcIDataElecInd}
\usage{
calcIDataElecInd()
}
\value{
OPENPROM input data IDataElecInd
}
\description{
Use ENERDATA electricity production data to derive OPENPROM input parameter IDataElecInd
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataElecInd", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
