% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataElecProd.R
\name{calcIDataElecProd}
\alias{calcIDataElecProd}
\title{calcIDataElecProd}
\usage{
calcIDataElecProd(mode)
}
\value{
OPENPROM input data iDataElecProd
}
\description{
Use ENERDATA electricity production data to derive OPENPROM input parameter iDataElecProd
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataElecProd", mode="NonCHP", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Michael Madianos
}
