% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataGrossInlCons.R
\name{calcIDataGrossInlCons}
\alias{calcIDataGrossInlCons}
\title{calcIDataGrossInlCons}
\usage{
calcIDataGrossInlCons()
}
\value{
OPENPROM input data
}
\description{
Use IEA data to derive OPENPROM input parameter IDataGrossInlCons
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataGrossInlCons", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
