% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDataHeatProd.R
\name{calcIDataHeatProd}
\alias{calcIDataHeatProd}
\title{calcIDataHeatProd}
\usage{
calcIDataHeatProd()
}
\value{
OPENPROM input data iDataHeatProd
}
\description{
Use IEA heat production data to derive OPENPROM input parameter iDataHeatProd
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataHeatProd", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
