% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataIntensityFertiliser.R
\name{calcIDataIntensityFertiliser}
\alias{calcIDataIntensityFertiliser}
\title{calcIDataIntensityFertiliser}
\usage{
calcIDataIntensityFertiliser()
}
\value{
Magpie object with the Food supply (kg/ha)
}
\description{
calcIDataIntensityFertiliser
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataIntensityFertiliser", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
