% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataOwnConsEne.R
\name{calcIDataOwnConsEne}
\alias{calcIDataOwnConsEne}
\title{calcIDataOwnConsEne}
\usage{
calcIDataOwnConsEne()
}
\value{
magpie object with OPENPROM input data IDataOwnConsEne
}
\description{
Use data from IEA to derive OPENPROM input parameter IDataOwnConsEne
This dataset includes own consumption values for each region and energy sector in Mtoe.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataOwnConsEne", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
