% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataPlantEffByType.R
\name{calcIDataPlantEffByType}
\alias{calcIDataPlantEffByType}
\title{calcIDataPlantEffByType}
\usage{
calcIDataPlantEffByType()
}
\value{
magpie object with OPENPROM input data iDataPlantEffByType.
}
\description{
Use data from EU Reference Scenario to derive OPENPROM input parameter iDataPlantEffByType
This dataset includes plant efficiency per plant type, as a ratio.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataPlantEffByType", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
