% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataScaleEndogScrap.R
\name{calcIDataScaleEndogScrap}
\alias{calcIDataScaleEndogScrap}
\title{calcIDataScaleEndogScrap}
\usage{
calcIDataScaleEndogScrap()
}
\value{
magpie object with OPENPROM input data iScaleEndogScrap
}
\description{
Use uncalibrated data to derive default values for iScaleEndogScrap
Itincludes parameters for premature scrapping in the industry module.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataScaleEndogScrap", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
