% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataShareBlend.R
\name{calcIDataShareBlend}
\alias{calcIDataShareBlend}
\title{calcIDataShareBlend}
\usage{
calcIDataShareBlend()
}
\value{
magpie object
}
\description{
Derive the fuel blend per demand subsectors for all countries
}
\examples{
\dontrun{
shareBlend <- calcOutput("IDataShareBlend", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
