% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataTechLftPlaType.R
\name{calcIDataTechLftPlaType}
\alias{calcIDataTechLftPlaType}
\title{calcIDataTechLftPlaType}
\usage{
calcIDataTechLftPlaType()
}
\value{
magpie object with OPENPROM input data iDataTechLftPlaType
}
\description{
Use data from EU Reference Scenario to derive OPENPROM input parameter iDataTechLftPlaType
This dataset includes technical lifetime per plant type, in years.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataTechLftPlaType", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
