% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataTrade.R
\name{calcIDataTrade}
\alias{calcIDataTrade}
\title{calcIDataTrade}
\usage{
calcIDataTrade()
}
\value{
magpie object with OPENPROM input data iDataTrade
}
\description{
Use data from IEA to derive OPENPROM input parameter iDataTrade
This dataset includes export/import values for each region and fuel type in Mtoe.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataTrade", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
