% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDataTransTech.R
\name{calcIDataTransTech}
\alias{calcIDataTransTech}
\title{calcIDataTransTech}
\usage{
calcIDataTransTech()
}
\value{
OPENPROM input data iDataTransTech
}
\description{
Use data to derive OPENPROM input parameter iDataTransTech
}
\examples{
\dontrun{
a <- calcOutput(type = "IDataTransTech", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
