% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIDecomPlants.R
\name{calcIDecomPlants}
\alias{calcIDecomPlants}
\title{calcIDecomPlants}
\usage{
calcIDecomPlants()
}
\value{
OPENPROM input data iIDecomPlants.
The output data calculated from the "Global Energy Monitor".
}
\description{
Use decommisioning capacities of power plants data from the Global Energy Monitor,
to derive OPENPROM input parameter iIDecomPlants.
}
\examples{
\dontrun{
a <- calcOutput(type = "IDecomPlants", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
