% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIEnvPolicies.R
\name{calcIEnvPolicies}
\alias{calcIEnvPolicies}
\title{calcIEnvPolicies}
\usage{
calcIEnvPolicies()
}
\value{
OPENPROM input data iEnvPolicies.
The output data when overlapping between EU Reference Scenario 2020 and
the ENGAGE project takes the EU Reference Scenario 2020 value.
}
\description{
Use carbon price data from the EU Reference Scenario 2020 and from the
ENGAGE project (carbon price data for the current policies scenario
GP_CurPol_T45) to derive OPENPROM input parameter iEnvPolicies.
}
\examples{
\dontrun{
a <- calcOutput(type = "IEnvPolicies", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
