% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIFixOandMCost.R
\name{calcIFixOandMCost}
\alias{calcIFixOandMCost}
\title{calcIFixOandMCost}
\usage{
calcIFixOandMCost()
}
\value{
magpie object with OPENPROM input data iFixOandMCost
}
\description{
Use data from EU Reference Scenario to derive OPENPROM input parameter iFixOandMCost
This dataset includes Fixed O&M costs per plant type, in EUR/kW.
}
\examples{
\dontrun{
a <- calcOutput(type = "IFixOandMCost", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Giannis Tolios, Fotis Sioutas
}
