% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIFuelCons.R
\name{calcIFuelCons}
\alias{calcIFuelCons}
\title{calcIFuelCons}
\usage{
calcIFuelCons(subtype = "TRANSE")
}
\arguments{
\item{subtype}{string, OPENPROM sector (DOMSE, INDSE, NENSE, TRANSE)}
}
\value{
OPENPROM input data iFuelConsXXX
}
\description{
Use ENERDATA, IEA, TREMOVE and NAVIGATE fuel consumption data to derive
OPENPROM input parameter iFuelConsXXX
The data for the years 2010 : 2021 is mainly from ENARDATA and IEA. 
For the years 2021:2100 the data is mainly from NAVIGATE.
For TRANSE the data from 2021:2100 the data is mainly from TREMOVE.
(XXX: NENSE, INDSE, DOMSE, TRANSE).
}
\examples{
\dontrun{
a <- calcOutput(type = "IFuelCons", subtype = "DOMSE", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
