% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIFuelCons2.R
\name{calcIFuelCons2}
\alias{calcIFuelCons2}
\title{calcIFuelCons2}
\usage{
calcIFuelCons2()
}
\arguments{
\item{subtype}{string, OPENPROM sector (DOMSE, INDSE, NENSE, TRANSE)}
}
\value{
OPENPROM input data iFuelConsXXX
}
\description{
Use ENERDATA, IEA, TREMOVE and NAVIGATE fuel consumption data to derive
OPENPROM input parameter iFuelConsXXX
The data for the years 2010 : 2021 is mainly from ENARDATA and IEA.
For the years 2021:2100 the data is mainly from NAVIGATE.
For TRANSE the data from 2021:2100 the data is mainly from TREMOVE.
(XXX: NENSE, INDSE, DOMSE, TRANSE).
}
\examples{
\dontrun{
a <- calcOutput(type = "IFuelCons2", subtype = "DOMSE", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
