% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIFuelPrice.R
\name{calcIFuelPrice}
\alias{calcIFuelPrice}
\title{calcIFuelPrice}
\usage{
calcIFuelPrice()
}
\value{
OPENPROM input data iFuelPrice
}
\description{
constant price from ENERDATA
Use ENERDATA fuel price data to derive OPENPROM input parameter iFuelPrice
}
\examples{
\dontrun{
a <- calcOutput(type = "IFuelPrice", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
