% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calciGDP.R
\name{calciGDP}
\alias{calciGDP}
\title{calciGDP}
\usage{
calciGDP(scenario = "SSP2")
}
\arguments{
\item{scenario}{string. By choosing a scenario you filter the SSP dataset
by type.}
}
\value{
The SSP data filtered by gdp
}
\description{
The SSP data filtered by gdp
}
\examples{
\dontrun{
gdp <- calcOutput("iGDP", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
