% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIGrossCapCosSubRen.R
\name{calcIGrossCapCosSubRen}
\alias{calcIGrossCapCosSubRen}
\title{calcIGrossCapCosSubRen}
\usage{
calcIGrossCapCosSubRen()
}
\value{
magpie object with OPENPROM input data iGrossCapCosSubRen.
}
\description{
Use data from EU Reference Scenario to derive OPENPROM input parameter iGrossCapCosSubRen
This dataset includes capital cost per plant type, in EUR/kW.
}
\examples{
\dontrun{
a <- calcOutput(type = "IGrossCapCosSubRen", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
