% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIH2Parameters.R
\name{calcIH2Parameters}
\alias{calcIH2Parameters}
\title{calcIH2Parameters}
\usage{
calcIH2Parameters()
}
\value{
OPENPROM input data IH2Parameters.
}
\description{
Use technology data from excel Common_DATA from MENA_EDS model
to derive OPENPROM input parameter IH2Parameters.
}
\examples{
\dontrun{
a <- calcOutput(type = "IH2Parameters", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
