% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIH2Production.R
\name{calcIH2Production}
\alias{calcIH2Production}
\title{calcIH2Production}
\usage{
calcIH2Production()
}
\value{
OPENPROM input data H2production.
The output data for Investment cost per unit of capacity (IC) from
technology cost and other data from the "EU Reference Scenario".
The output data for Fixed Costs (FC) and (EFF)  is from "EU Reference Scenario".
The output data for Variable Costs (VC), (AVAIL), (LFT) and (CR) is from excel
Common_DATA from MENA_EDS model.
}
\description{
Use technology cost data from the "EU Reference Scenario", MENA_EDS model data,
Technical Lifetime, CR and AVAIL data from various sources to derive OPENPROM
input parameter IH2Production.
}
\examples{
\dontrun{
a <- calcOutput(type = "IH2Production", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
