% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIInstCapPast.R
\name{calcIInstCapPast}
\alias{calcIInstCapPast}
\title{calcIInstCapPast}
\usage{
calcIInstCapPast()
}
\value{
OPENPROM input data iInstCapPast
}
\description{
Use data to derive OPENPROM input parameter iInstCapPast
}
\examples{
\dontrun{
a <- calcOutput(type = "IInstCapPast", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
