% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIInstCapPast.R
\name{calcIInstCapPast}
\alias{calcIInstCapPast}
\title{calcIInstCapPast}
\usage{
calcIInstCapPast(mode = "TotalEff")
}
\value{
OPENPROM input data iInstCapPast
}
\description{
Use data from ENERDATA to derive OPENPROM input parameter iInstCapPast
This dataset contains the values of installed capacity for past years in GW.
}
\examples{
\dontrun{
a <- calcOutput(type = "IInstCapPast", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios, Michael Madianos
}
