% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIInstCapPast2.R
\name{calcIInstCapPast2}
\alias{calcIInstCapPast2}
\title{calcIInstCapPast2}
\usage{
calcIInstCapPast2(mode = "TotalEff")
}
\value{
OPENPROM input data iInstCapPast
}
\description{
Use data from IRENA to derive OPENPROM input parameter iInstCapPast2
This dataset contains the values of installed capacity for past years in GW.
}
\examples{
\dontrun{
a <- calcOutput(type = "IInstCapPast2", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios, Michael Madianos
}
