% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIMatFacPlaAvailCap.R
\name{calcIMatFacPlaAvailCap}
\alias{calcIMatFacPlaAvailCap}
\title{calcIMatFacPlaAvailCap}
\usage{
calcIMatFacPlaAvailCap()
}
\value{
magpie object with OPENPROM input data iMatFacPlaAvailCap.
}
\description{
Use uncalibrated data to derive default values for iMatFacPlaAvailCap.
This dataset includes Maturity factors on PG Capacity.
}
\examples{
\dontrun{
a <- calcOutput(type = "IMatFacPlaAvailCap", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
