% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcINewREg.R
\name{calcINewREg}
\alias{calcINewREg}
\title{calcINewREg}
\usage{
calcINewREg()
}
\value{
Magpie object with passenger-cars-in-use per year
}
\description{
Calculate the passenger-cars-in-use per year
}
\examples{
\dontrun{
a <- calcOutput(type = "INewREg", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
