% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcINewReg.R
\name{calcINewReg}
\alias{calcINewReg}
\title{calcINewReg}
\usage{
calcINewReg()
}
\value{
Magpie object with the passenger-car-first-registrations per year
}
\description{
Calculate the passenger-car-first-registrations per year
}
\examples{
\dontrun{
a <- calcOutput(type = "INewReg", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
