% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIPriceFuelsInt.R
\name{calcIPriceFuelsInt}
\alias{calcIPriceFuelsInt}
\title{calcIPriceFuelsInt}
\usage{
calcIPriceFuelsInt()
}
\value{
magpie object with OPENPROM input data iPriceFuelsInt.
}
\description{
Use data from IEA and ENERDATA to derive OPENPROM input parameter iPriceFuelsInt
This dataset includes fossil fuel prices from IEA World Energy Outlook and ENERDATA, in $2015/toe.
}
\examples{
\dontrun{
a <- calcOutput(type = "IPriceFuelsInt", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
