% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIPriceFuelsIntBase.R
\name{calcIPriceFuelsIntBase}
\alias{calcIPriceFuelsIntBase}
\title{calcIPriceFuelsIntBase}
\usage{
calcIPriceFuelsIntBase()
}
\value{
magpie object with OPENPROM input data iPriceFuelsIntBase.
}
\description{
Use data from IEA and ENERDATA to derive OPENPROM input parameter iPriceFuelsIntBase
This dataset includes fossil fuel prices from IEA World Energy Outlook 2023, in $2015/toe.
}
\examples{
\dontrun{
a <- calcOutput(type = "IPriceFuelsIntBase", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
