% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIRateLossesFinCons.R
\name{calcIRateLossesFinCons}
\alias{calcIRateLossesFinCons}
\title{calcIRateLossesFinCons}
\usage{
calcIRateLossesFinCons()
}
\value{
OPENPROM input data iRateLossesFinCons
}
\description{
Use enerdata and IRF data to derive OPENPROM input parameter iRateLossesFinCons
(distribution losses over final consumption)
}
\examples{
\dontrun{
a <- calcOutput(type = "IRateLossesFinCons", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
