% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIRatioBranchOwnCons.R
\name{calcIRatioBranchOwnCons}
\alias{calcIRatioBranchOwnCons}
\title{calcIRatioBranchOwnCons}
\usage{
calcIRatioBranchOwnCons()
}
\value{
OPENPROM input data iRatioBranchOwnCons.
The output data calculated from the IEA.
}
\description{
Use ...
to derive OPENPROM input parameter iRatioBranchOwnCons.
}
\examples{
\dontrun{
a <- calcOutput(type = "IRatioBranchOwnCons", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
