% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISFC.R
\name{calcISFC}
\alias{calcISFC}
\title{calcISFC}
\usage{
calcISFC()
}
\value{
magpie object
}
\description{
Derive Specific fuel consumption per car technology for all countries
}
\examples{
\dontrun{
SFC <- calcOutput("ISFC", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
