% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISubsiPerDemTech.R
\name{calcISubsiPerDemTech}
\alias{calcISubsiPerDemTech}
\title{calcISubsiPerDemTech}
\usage{
calcISubsiPerDemTech()
}
\value{
OPENPROM input data iSubsiPerDemTech
}
\description{
Use data to derive OPENPROM input parameter iSubsiPerDemTech
}
\examples{
\dontrun{
a <- calcOutput(type = "ISubsiPerDemTech", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
