% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISuppRatePrimProd.R
\name{calcISuppRatePrimProd}
\alias{calcISuppRatePrimProd}
\title{calcISuppRatePrimProd}
\usage{
calcISuppRatePrimProd()
}
\value{
OPENPROM input data iSuppRatePrimProd.
}
\description{
Use ENERDATA data to derive OPENPROM input parameter iSuppRatePrimProd.
Primary production data / sum of FUEL
}
\examples{
\dontrun{
a <- calcOutput(type = "ISuppRatePrimProd", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
