% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISuppTransfers.R
\name{calcISuppTransfers}
\alias{calcISuppTransfers}
\title{calcISuppTransfers}
\usage{
calcISuppTransfers()
}
\value{
magpie object with OPENPROM input data iSuppTransfers.
}
\description{
Use data from IEA to derive OPENPROM input parameter iSuppTransfers
This dataset includes the supplementary parameter for transfers, in Mtoe.
}
\examples{
\dontrun{
a <- calcOutput(type = "ISuppTransfers", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
