% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcITotEneSupply.R
\name{calcITotEneSupply}
\alias{calcITotEneSupply}
\title{calcITotEneSupply}
\usage{
calcITotEneSupply(subtype)
}
\value{
OPENPROM input data IPrimProd
}
\description{
Use IEA Total Energy Supply data to derive OPENPROM input parameter IPrimProd
}
\examples{
\dontrun{
a <- calcOutput(type = "ITotEneSupply", subtype = "Primary", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
