% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcITransChar.R
\name{calcITransChar}
\alias{calcITransChar}
\title{calcITransChar}
\usage{
calcITransChar()
}
\value{
OPENPROM input data iTransChar
}
\description{
Use IRF data to derive OPENPROM input parameter iTransChar
km per car, passengers per car and residuals for passenger cars market extension
}
\examples{
\dontrun{
a <- calcOutput(type = "ITransChar", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
