% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcITransfProcesses.R
\name{calcITransfProcess}
\alias{calcITransfProcess}
\title{calcITransfProcess}
\usage{
calcITransfProcess(subtype = "Total", flow = "Inp")
}
\value{
OPENPROM input data ITransfProcess.
}
\description{
Use IEA transformation processes data to derive OPENPROM input parameter ITransfProcess.
}
\examples{
\dontrun{
a <- calcOutput(type = "ITransfProcess", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
