% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIVarCost.R
\name{calcIVarCost}
\alias{calcIVarCost}
\title{calcIVarCost}
\usage{
calcIVarCost()
}
\value{
magpie object with OPENPROM input data iVarCost
}
\description{
Use data from EU Reference Scenario to derive OPENPROM input parameter iVarCost
This dataset includes variable non fuel cost per plant type, in EUR/MWh.
}
\examples{
\dontrun{
a <- calcOutput(type = "IVarCost", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios
}
