% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNavigate.R
\name{calcNavigate}
\alias{calcNavigate}
\title{calcNavigate}
\usage{
calcNavigate(subtype = "DOMSE")
}
\arguments{
\item{subtype}{string sector (DOMSE, INDSE, NENSE, TRANSE)}
}
\value{
Navigate and calcIFuelCons fuel consumption in XXX sector
}
\description{
Use Navigate and calcIFuelCons fuel consumption in XXX sector.
The data for the years 2010 : 2021 is from calcIFuelCons.
For TRANSE is until year 2020 from calcIFuelCons.
The data for the years 2022 : 2100 is from Navigate.
(XXX: NENSE, INDSE, DOMSE, TRANSE).
}
\examples{
\dontrun{
a <- calcOutput(type = "Navigate", subtype = "DOMSE", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
