% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNavigateEmissions.R
\name{calcNavigateEmissions}
\alias{calcNavigateEmissions}
\title{calcNavigateEmissions}
\usage{
calcNavigateEmissions()
}
\value{
Navigate Emissions as magpie object
}
\description{
Use Navigate Emissions to reporting.
}
\examples{
\dontrun{
a <- calcOutput(type = "NavigateEmissions", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
