% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPOP.R
\name{calcPOP}
\alias{calcPOP}
\title{calcPOP}
\usage{
calcPOP(scenario = "SSP2")
}
\arguments{
\item{scenario}{string. By choosing a scenario you filter the SSP dataset
by type.}
}
\value{
The SSP data filtered by pop
}
\description{
The SSP data filtered by pop
}
\examples{
\dontrun{
POP <- calcOutput("POP", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
