% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrimes.R
\name{calcPrimes}
\alias{calcPrimes}
\title{calcPrimes}
\usage{
calcPrimes()
}
\value{
Primes fuel consumption in DOMSE, INDSE, NENSE, TRANSE
}
\description{
Use Primes data for fuel consumption in DOMSE, INDSE, NENSE, TRANSE
}
\examples{
\dontrun{
a <- calcOutput(type = "Primes", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
