% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPrimesPrices.R
\name{calcPrimesPrices}
\alias{calcPrimesPrices}
\title{calcPrimesPrices}
\usage{
calcPrimesPrices()
}
\value{
Primes Prices data for END USER PRICE
}
\description{
Use Primes Prices data for END USER PRICE
}
\examples{
\dontrun{
a <- calcOutput(type = "PrimesPrices", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
