% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStockPC.R
\name{calcStockPC}
\alias{calcStockPC}
\title{calcStockPC}
\usage{
calcStockPC()
}
\value{
magpie object
}
\description{
Derive car stock per fuel technology for all countries.
}
\examples{
\dontrun{
stockPC <- calcOutput("StockPC", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
