% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTDemand.R
\name{calcTDemand}
\alias{calcTDemand}
\title{calcTDemand}
\usage{
calcTDemand()
}
\value{
Secondary Energy Electricity
}
\description{
Use Primes, IEA and Navigate data for Secondary Energy Electricity
}
\examples{
\dontrun{
a <- calcOutput(type = "TDemand", aggregate = FALSE)
}

}
\author{
Fotis Sioutas
}
