% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTFuelCons.R
\name{calcTFuelCons}
\alias{calcTFuelCons}
\title{calcTFuelCons}
\usage{
calcTFuelCons()
}
\value{
OPENPROM input targets iFuelCons
}
\description{
calcTFuelCons
}
\examples{
\dontrun{
a <- calcOutput(type = "TFuelCons")
}

}
\author{
Fotis Sioutas
}
