% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTFuelConsShares.R
\name{calcTFuelConsShares}
\alias{calcTFuelConsShares}
\title{calcTFuelConsShares}
\usage{
calcTFuelConsShares(subtype = "TRANSE")
}
\value{
OPENPROM input data iFuelConsXXX
}
\description{
Fuel consumption share projections
Aggregates projected shares of fuel consumption by fuel type from multiple
sectoral projection sources (domestic, PCH, NEN  and industrial sectors) to the
country (ISO3) level. Aggregation is performed using fuel-consumption-based
weights derived from historical fuel consumption data.
}
\examples{
\dontrun{
a <- calcOutput(type = "TFuelConsShares")
}

}
\author{
Fotis Sioutas
}
