% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTInstCap.R
\name{calcTInstCap}
\alias{calcTInstCap}
\title{calcTInstCap}
\usage{
calcTInstCap()
}
\value{
magpie object
}
\description{
Use capacities to generate targets for capacity.
}
\examples{
\dontrun{
a <- calcOutput(type = "TInstCap", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
