% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTNewShareStockPC.R
\name{calcTNewShareStockPC}
\alias{calcTNewShareStockPC}
\title{calcTNewShareStockPC}
\usage{
calcTNewShareStockPC()
}
\arguments{
\item{subtype}{Period that starts the projection.}
}
\value{
magpie object
}
\description{
Derive car stock per fuel technology for all countries.
}
\examples{
\dontrun{
TStockPC <- calcOutput("NewShareStockPC", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
