% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTProdElec.R
\name{calcTProdElec}
\alias{calcTProdElec}
\title{calcTProdElec}
\usage{
calcTProdElec()
}
\value{
magpie object
}
\description{
Use ProdElec to generate targets for ProdElec
}
\details{
Info:
IEA: calciDataProdElec data, shares for data that is missing from ENERDATA, until 2024
Primes: Primes data, shares for data that is missing from ENERDATA, EU countries until 2070,multiply by IEA trends(after 2070).
IEA: IEA ProdElec data, find trends for ProdElec for each year.
The trends are the same for each country depending to the region. For example
HKG and CHN have the same trends for ProdElec
Shares for data that is missing from ENERDATA, 225 countries until 2050.
#' IEA mapping: "Africa" = "SSA", "Middle East" = "MEA", "Eurasia" = "REF",
"Southeast Asia" = "OAS", "Central and South America" = "LAM",
"Asia Pacific" = "CAZ", "Europe" = "NEU", "European Union" = "ELL"
calculate CAZ, NEU and ELL 
"CAZ" <- "CAZ" -  "OAS"
ELL and NEU have the same trends
IEA_non_EU <- "NEU" - "ELL"
"NEU" <- IEA_non_EU
"ELL" <- IEA_non_EU
The trends are multiplied with the historical from calciDataProdElec IEA data to find the ProdElec.
}
\examples{
\dontrun{
a <- calcOutput(type = "TProdElec", aggregate = FALSE)
}

}
\author{
Michael Madianos
}
