% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTREMOVE.R
\name{calcTREMOVE}
\alias{calcTREMOVE}
\title{calcTREMOVE}
\usage{
calcTREMOVE()
}
\value{
TREMOVE fuel consumption in TRANSE sector
}
\description{
Use TREMOVE data for fuel consumption in TRANSE sector
}
\examples{
\dontrun{
a <- calcOutput(type = "TREMOVE", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
