% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTStockPC.R
\name{calcTStockPC}
\alias{calcTStockPC}
\title{calcTStockPC}
\usage{
calcTStockPC(subtype = 2021)
}
\arguments{
\item{subtype}{Period that starts the projection.}
}
\value{
magpie object
}
\description{
Derive car stock per fuel technology for all countries.
}
\examples{
\dontrun{
TStockPC <- calcOutput("TStockPC", aggregate = FALSE)
}

}
\author{
Michael Madianos, Anastasis Giannousakis
}
