% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcUNFCCC_IP_CO2.R
\name{calcUNFCCC_IP_CO2}
\alias{calcUNFCCC_IP_CO2}
\title{calcUNFCCC_IP_CO2}
\usage{
calcUNFCCC_IP_CO2()
}
\value{
OPENPROM input data UNFCCC_IP_CO2
}
\description{
CO2 emissions from industrial processes "Mineral industry,
Chemical industry, Metal industry" from UNFCCC emissions data.
To derive OPENPROM input parameter UNFCCC_IP_CO2.
The data is from the "UNFCCC".
}
\examples{
\dontrun{
a <- calcOutput(type = "UNFCCC_IP_CO2", aggregate = FALSE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Giannis Tolios, Maro Baka
}
