% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertENERDATA.R
\name{convertENERDATA}
\alias{convertENERDATA}
\title{convertENERDATA}
\usage{
convertENERDATA(x)
}
\arguments{
\item{x}{MAgPIE object with ISO country codes.}
}
\value{
The "ENERDATA" data with spatial entries for each country.
}
\description{
NA values are replaced with zeros and historical data are corrected.
}
\author{
Anastasis Giannousakis, Fotis Sioutas

#' @examples
\dontrun{
a <- readSource("ENERDATA", convert = TRUE)
}
}
