% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertENERDATA.R
\name{convertENERDATA}
\alias{convertENERDATA}
\title{convertENERDATA}
\usage{
convertENERDATA(x, subtype)
}
\arguments{
\item{x}{MAgPIE object with ISO country codes.}

\item{subtype}{string. By choosing a subtype you filter the ENERDATA dataset
(1800+ variables) by type, to allow further processing of specific variables}
}
\value{
The "ENERDATA" data (filtered by subtype).
}
\description{
The two available ENERDATA datasets are combined into one,
NA values are replaced with zeros and historical data are corrected.
}
\examples{
\dontrun{
a <- readSource("ENERDATA", convert = TRUE)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
