% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullHISTORICAL.R
\name{fullHISTORICAL}
\alias{fullHISTORICAL}
\title{fullHISTORICAL}
\usage{
fullHISTORICAL()
}
\value{
The mif file
}
\description{
Read data, convert it to a mrprom mif file so to compare output mif file
with OPEN-PROM output.
}
\examples{
\dontrun{
a <- retrieveData("HISTORICAL", regionmapping = "regionmappingOP.csv")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
