% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullHISTORICAL.R
\name{fullHISTORICAL}
\alias{fullHISTORICAL}
\title{fullHISTORICAL}
\usage{
fullHISTORICAL()
}
\value{
The mif file
}
\description{
Read files, convert it to a mif file so to compare mrprom output mif file
with OPEN-PROM output.
}
\examples{
\dontrun{
a <- retrieveData("Historical", regionmapping = "regionmappingOPDEV2.csv")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
