% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullVALIDATION.R
\name{fullVALIDATION}
\alias{fullVALIDATION}
\title{fullVALIDATION}
\usage{
fullVALIDATION()
}
\value{
The mif file
}
\description{
Read data from MENA-EDS, ENERDATA, IEA and NAVIGATE convert it to a mrprom mif
file so to compare output mif file with OPEN-PROM output.
}
\examples{
\dontrun{
a <- retrieveData("VALIDATION", regionmapping = "regionmappingOPDEV3.csv")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
