% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modellingTask.R
\name{modellingTask}
\alias{modellingTask}
\title{modellingTask}
\usage{
modellingTask(file)
}
\arguments{
\item{file}{The input.gms file path.}
}
\value{
Info about sources of mrprom.
}
\description{
Description of the sources of mrprom.
}
\examples{
\dontrun{
a <- modellingTask(file)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
