% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readClimatePolicyModelling.R
\name{readClimatePolicyModelling}
\alias{readClimatePolicyModelling}
\title{Read Climate Policy Modelling data}
\usage{
readClimatePolicyModelling(subtype)
}
\arguments{
\item{subtype}{Character. Available subtypes are:
\itemize{
  \item `High impact policies_update`
  \item `Translation to model target`
  \item `Archived policies`
}}
}
\value{
A list containing the data as a quitte object and its metadata.
}
\description{
Reads a selected sheet from the Climate Policy Modelling Protocol workbook,
removes rows marked as deleted using strikethrough formatting, and converts
the result into a quitte object.
}
\examples{
\dontrun{
a <- readSource("ClimatePolicyModelling",subtype = "Translation to model target")
}

}
\author{
Fotis Sioutas
}
