% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readENERDATA.R
\name{readENERDATA}
\alias{readENERDATA}
\title{ENERDATA}
\usage{
readENERDATA(subtype = "capacity")
}
\arguments{
\item{subtype}{The type is referring to the capacity of electricity and the
data that should be read is from the csv file "open_prom_database_20_4_2023_1.csv"
and convert it to a magpie object.}
}
\value{
The read-in data into a magpie object
}
\description{
Read in an csv file and convert it to a magpie object
The data has information about electricity capacity per country and per year
}
\author{
Anastasis Giannousakis, Fotis Sioutas

#' @examples
\dontrun{
a <- readSource("ENERDATA")
}
}
