% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readENERDATA2.R
\name{readENERDATA2}
\alias{readENERDATA2}
\title{readENERDATA2}
\usage{
readENERDATA2()
}
\value{
The read-in data into a magpie object
}
\description{
Read a csv files with data from ENERDATA.
The dataset contains several energy-related data types from ENERDATA for
various countries and years.
}
\examples{
\dontrun{
a <- readSource("ENERDATA")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
