% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readff55NonCo2GHGEmissions.R
\name{readFf55NonCo2GhgEmissions}
\alias{readFf55NonCo2GhgEmissions}
\title{readff55NonCo2GHGEmissions}
\usage{
readFf55NonCo2GhgEmissions()
}
\value{
The read-in data into a magpie object.
}
\description{
Read ff55NonCo2GHGEmissions of EU
}
\examples{
\dontrun{
a <- readSource("ff55NonCo2GHGEmissions", convert = FALSE)
}

}
\author{
Fotis Sioutas
}
