% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEME3.R
\name{readGEME3}
\alias{readGEME3}
\title{readGEME3}
\usage{
readGEME3(subtype = "Npi")
}
\arguments{
\item{subtype}{Type of carbon prices}
}
\value{
The read-in data into a magpie object.
}
\description{
Read Production Level and Unit Cost data as delivered in GDX files from GEME3.
}
\examples{
\dontrun{
a <- readSource("GEME3", subtype = "Npi")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
