% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_Industry_Roadmaps.R
\name{readIEA_Industry_Roadmaps}
\alias{readIEA_Industry_Roadmaps}
\title{readIEA_Industry_Roadmaps}
\usage{
readIEA_Industry_Roadmaps(subtype = "IEA_IIS_Roadmap")
}
\arguments{
\item{subtype}{Type of data that should be read. The type is referring to the
excel sheet, from the excel file "IEA_Industries_Roadmaps_Reports.xlsx"
and convert it to a magpie object.
Available types are:
\itemize{
\item `IEA_IIS_Roadmap`:
\item `IEA_Tech_Assumptions`:
}}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in Regional energy demand for steelmaking and EAD shares and tech shares
}
\examples{
\dontrun{
a <- readSource("IEA_Industry_Roadmaps")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas, Sonja Sechi
}
