% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIRF.R
\name{readIRF}
\alias{readIRF}
\title{readIRF}
\usage{
readIRF(subtype = "passenger-cars-in-use")
}
\arguments{
\item{subtype}{By choosing a subtype you can select the corresponding csv
file and convert it to a magpie object.
Available types are:
\itemize{
\item `passenger-cars-in-use`:
\item `buses-and-motor-coaches-in-use`:
\item `total-vans,-pickups,-lorries-and-road-tractors-in-use`:
\item `bus-and-motor-coach-traffic`:
\item `lorry-and-road-tractor-traffic`:
\item `inland-surface-freight-transport-by-inland-waterway`:
\item `inland-surface-freight-transport-by-rail`:
\item `inland-surface-freight-transport-by-road`:
\item `inland-surface-passenger-transport-by-rail`:
\item `inland-surface-public-passenger-transport-by-road`:
\item `passenger-car-traffic`:
\item `total-van,-pickup,-lorry-and-road-tractor-traffic`:
\item `passenger-car-first-registrations`:
\item `total-vehicles-in-use`:
}}
}
\value{
The read-in data into a magpie object
}
\description{
Read in an csv file and convert it to a magpie object.
The dataset contains several data types about transport and traffic.
The data is from https://datawarehouse.worldroadstatistics.org/users/login
The International Road Federation serves a network of public and private
sector members by providing world-class knowledge resources,
advocacy services, and continuing education programs which together offer a
global marketplace for best practices and industry solutions"
}
\examples{
\dontrun{
a <- readSource("IRF")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
