% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNavigate.R
\name{readNavigate}
\alias{readNavigate}
\title{readNavigate}
\usage{
readNavigate(subtype = "SUP_NPi_Default")
}
\arguments{
\item{subtype}{Type of data that should be read.}
}
\value{
The read-in data into a magpie object.
}
\description{
Read in NAVIGATE project model scenarios from IIASA.
IAM results :  Final Energy, Primary Production,
Secondary Energy Electricity, Emissions CO2
The models are : 
COFFEE 1.5, REMIND-MAgPIE 3.2-4.6, IMAGE 3.3,
WITCH 5.0, IMACLIM 2.0, MESSAGEix-Materials.
}
\details{
The scenarios are : 
SUP_NPi_Default,SUP_2C_Default,SUP_1p5C_Default",
SUP_2C_Regional,SUP_1p5C_Regional,SUP_1p5C_Elec,SUP_1p5C_Elec_LimNuc,
SUP_1p5C_Elec_LimCCS,SUP_1p5C_Elec_HighVRE,SUP_1p5C_Elec_OptMAC,
SUP_1p5C_Comb,SUP_1p5C_Comb_LimNuc,SUP_1p5C_Comb_LimCCS",
SUP_1p5C_Comb_HighVRE,SUP_1p5C_Comb_OptMAC,SUP_2C_Elec,
SUP_2C_Elec_LimNuc,SUP_2C_Elec_LimCCS,SUP_2C_Elec_HighVRE,
SUP_2C_Elec_OptMAC,SUP_2C_Comb,SUP_2C_Comb_LimNuc,
SUP_2C_Comb_LimCCS,SUP_2C_Comb_HighVRE,SUP_2C_Comb_OptMAC,
NAV_Dem-NPi-ref,NAV_Ind_NPi
}
\examples{
\dontrun{
a <- readSource("Navigate", subtype = "SUP_NPi_Default")
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
