% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrimesPGData.R
\name{readPrimesPGData}
\alias{readPrimesPGData}
\title{readPrimesPGData}
\usage{
readPrimesPGData(subtype = "SE")
}
\arguments{
\item{subtype}{excel sheet "SE" for SE Electricity or "power generation" for
Gross Electricity generation by plant type and "capacity" for
Net Installed Power Capacity}
}
\value{
The read-in data into a magpie object.
}
\description{
Read Primes Secondary Energy Electricity Gross Electricity generation by plant type
and Net Installed Power Capacity for EU countries:
}
\examples{
\dontrun{
a <- readSource("PrimesPGData", subtype = "SE")
}

}
\author{
Fotis Sioutas
}
