% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTechCosts.R
\name{readTechCosts}
\alias{readTechCosts}
\title{readTechCosts}
\usage{
readTechCosts(subtype = "PowerAndHeat")
}
\arguments{
\item{subtype}{Type of data that should be read. The type is referring to the excel sheet, from the excel
file "REF2020_Technology Assumptions_Transport.xlsx" and convert it to a magpie object.
Available types are:
\itemize{
\item `PowerAndHeat`:
\item `DomesticEnergy`:
\item `IndustryEnergy`:
\item `infrastructure`:
\item `new_fuels_energy`:
\item `maritime`:
\item `Inland_navigation`:
\item `Rail`:
\item `Aviation`:
\item `2wheelers`:
\item `Bus_coach`:
\item `HGVs>16t`:
\item `HGVs<16t`:
\item `LCVs`:
\item `Large_cars`:
\item `Medium_cars`:
\item `Small_cars`:
\item `renovation_costs`:
}}
}
\value{
The read-in data into a magpie object
}
\description{
Read in an excel file and convert it to a magpie object
The file containing the data is named TechCosts
}
\author{
Anastasis Giannousakis, Fotis Sioutas

#' @examples
\dontrun{
a <- readSource("TechCosts")
}
}
