% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUPTCarbonPrices.R
\name{readUPTCarbonPrices}
\alias{readUPTCarbonPrices}
\title{readUPTCarbonPrices}
\usage{
readUPTCarbonPrices()
}
\value{
The read-in carbon price data into a magpie object
}
\description{
Read UPT Carbon Prices
The dataset contains carbon price data.
}
\examples{
\dontrun{
a <- readSource("UPTCarbonPrices")
}

}
\author{
Fotis Sioutas
}
