% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFunnelPlots.R
\name{toolFunnelPlots}
\alias{toolFunnelPlots}
\title{toolFunnelPlots}
\usage{
toolFunnelPlots()
}
\arguments{
\item{subtypes}{magpie object}
}
\value{
The read-in data to plots
}
\description{
Make FunnelPlots from magpie object for specific regions and variables
with many models and scenarios and years.
}
\examples{
\dontrun{
a <- toolFunnelPlots(magpie_object = x)
}

}
\author{
Anastasis Giannousakis, Fotis Sioutas
}
