% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetMissingCountries.R
\name{toolGetMissingCountries}
\alias{toolGetMissingCountries}
\title{toolGetMissingCountries}
\usage{
toolGetMissingCountries(x)
}
\arguments{
\item{x}{Vector of ISO3 country codes}
}
\value{
A vector of missing country names
}
\description{
For a given vector of ISO3 country codes, this function returns
the full names of countries that are in the official countries list
and are missing from the given vector
}
\examples{
\dontrun{
a <- toolGetMissingCountries(c("DEU", "USA"))
}


}
\author{
Anastasis Giannousakis
}
