% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReadSetFromGDX.R
\name{toolReadSetFromGDX}
\alias{toolReadSetFromGDX}
\title{toolReadSetFromGDX}
\usage{
toolReadSetFromGDX(gdx_file = "fulldata", set = "SECTTECH")
}
\arguments{
\item{gdx_file}{Name of the gdx file}

\item{set}{sets to read from gdx file}
}
\value{
Dataframe with the values_parameters of some variables
from the gdx file.
}
\description{
For the given gdx file, this function returns a dataframe with the
values_parameters of some variables from the gdx file.
}
\examples{
\dontrun{
a <- toolReadSetFromGDX(gdx_file = "fulldata", set = c("SECTTECH", "INDSE")
}

}
\author{
Anastasis Giannousakis Fotis Sioutas
}
