% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSet.R
\name{toolSet}
\alias{toolSet}
\title{toolSet}
\usage{
toolSet(x, set, type)
}
\arguments{
\item{x}{dataframe}

\item{set}{column of (x)}

\item{type}{type of data.}
}
\value{
A list in order to create a gdx file.
}
\description{
For a given dataframe (x) containing a column called set, 
this function returns a list with the unique values of the column set
and with the standard format in order to create a gdx file.
}
\examples{
\dontrun{
a <- toolSet(x, "region","IEA")
}


}
\author{
Anastasis Giannousakis Fotis Sioutas
}
