% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSubtype.R
\name{toolSubtype}
\alias{toolSubtype}
\title{toolSubtype}
\usage{
toolSubtype(x, subt, type)
}
\arguments{
\item{x}{dataframe}

\item{subt}{string. By choosing a subtype you filter the dataset.}

\item{type}{type of data.}
}
\value{
A list in order to create a gdx file.
}
\description{
For a given dataframe (x), this function returns
a list with the values_parameters of (x) and the unique values
of each column of (x) with the standard format
in order to create a gdx file.
}
\examples{
\dontrun{
a <- toolSubtype(x, "nama_10_pe", "Eurostat")
}

}
\author{
Anastasis Giannousakis Fotis Sioutas
}
